### 2020-07-02 #25 @pmeier
pystiche: A Framework for Neural Style Transfer
Submitting Author: Philip Meier @pmeier
Package Name: `pystiche`
One-Line Description of Package: Framework for Neural Style Transfer (NST) built upon PyTorch
Repository Link:  https://github.com/pmeier/pystiche
Version submitted: [`pystiche==0.5.0`](https://github.com/pmeier/pystiche/tree/v0.5.0)
Version accepted: TBD

---

`pystiche` is a framework for Neural Style Transfer (NST) algorithms based on [PyTorch](https://pytorch.org). NST is a neural-net-based technique to [merge the content of one and the artistic style of another image](https://pytorch.org/tutorials/advanced/neural_style_tutorial.html). Similar to deep learning frameworks `pystiche` eases up the workflow for researchers in this field. Rather than implementing everything yourself, `pystiche` provides common building blocks of NST algorithms that can be conveniently combined. Thus, researchers can focus on implementing new ideas rather than implementing the periphery over and over again.

- Please indicate which [category or categories](https://www.pyopensci.org/dev_guide/peer_review/aims_scope.html) this package falls under:

- Explain how the and why the package falls under these categories (briefly, 1-2 sentences):

  `pystiche` can be used to reproduce NST papers while focusing on core aspects.

  The primary intended audience are researchers as described above. Apart from them `pystiche` could also be interesting for recreational use by non-scientists.

  AFAIK there are no other packages provide a similar functionality. However, due to its popularity, there are many implementations, which are limited to a specific NST algorithm. An exception might be [this](https://github.com/sunshineatnoon/Paper-Implementations/tree/master/NeuralSytleTransfer) which features the implementation of multiple papers.

    #21

For details about the pyOpenSci packaging requirements, see our [packaging guide](https://www.pyopensci.org/dev_guide/packaging/packaging_guide.html). Confirm each of the following by checking the box.  This package:

- [x] has an [OSI approved license](https://opensource.org/licenses)

- [x] The package has an **obvious research application** according to JOSS's definition in their [submission requirements](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements). Be aware that completing the pyOpenSci review process **does not** guarantee acceptance to JOSS. Be sure to read their submission requirements (linked above) if you are interested in submitting to JOSS.
- [x] The package is not a "minor utility" as defined by JOSS's [submission requirements](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements): "Minor ‘utility’ packages, including ‘thin’ API clients, are not acceptable." pyOpenSci welcomes these packages under "Data Retrieval", but JOSS has slightly different criteria.

- [x] I agree to abide by [pyOpenSci's Code of Conduct](https://www.pyopensci.org/contributing-guide/peer_review/coc.html) during the review process and in maintaining my package should it be accepted.

**P.S.** *Have feedback/comments about our review process? Leave a comment [here](https://github.com/pyOpenSci/governance/issues/8)*

[Editor and review templates can be found here](https://www.pyopensci.org/contributing-guide/appendices/templates.html)
