### 2020-07-02 #25 @pmeier
Some addtional remarks:

- I'm currently battling with [Zenodo](https://help.zenodo.org/) since the webhook did not work when I created the `v0.5.0` release. To make matters worse the manual upload is also not working for. I'll work this out soon.
- As discussed in https://github.com/pyOpenSci/software-review/issues/21#issuecomment-626795429 (@lwasser, @NickleDave) I've found two reviewers for the technical aspects of `pystiche`:

  - @edgarriba: Edgar Riba is the creator of [`kornia`](https://kornia.github.io/), a CV library for PyTorch.
  - @soumith: Soumith Chintala is a member of the [PyTorch core team](https://pytorch.org/docs/stable/community/persons_of_interest.html#general-maintainers).

  My apologies for the late submission. Due to CoViD I had to help out in other more pressing projects first. I hope your willingness to review still stands.
