### 2020-07-02 7dd1628
bump 0.5.0 to 1.0.0+dev (#335)
