### 2020-07-13 72e3b0a
pin isort < 5 (#337)
