+++ requirements-lint.txt
@@ -1,4 +1,4 @@
-isort
+isort < 5
 black
 flake8 >= 3.8
 mypy
