### 2020-07-13 3854d2e
switch from TravisCI and AppVeyor to GitHub Actions (#329)
* switch from TravisCI and AppVeyor to GitHub Actions

* remove obsolete CI configs and scripts
