+++ .ci/after_success.sh
@@ -1,4 +1,0 @@
-#!/usr/bin/env sh
-
-coverage report
-codecov -X gcov
