+++ appveyor.yml
@@ -1,29 +1,0 @@
-branches:
-  only:
-    - master
-
-matrix:
-  fast_finish: true
-
-environment:
-  matrix:
-    - PYTHON_DIR: "C:\\Python36-x64"
-    - PYTHON_DIR: "C:\\Python37-x64"
-    - PYTHON_DIR: "C:\\Python38-x64"
-
-init:
-  - set PATH=%PYTHON_DIR%;%PYTHON_DIR%\Scripts;%PATH%
-  - mklink %PYTHON_DIR%\python3.exe %PYTHON_DIR%\python.exe
-  - python3 --version
-  - pip3 --version
-
-build: false
-
-install:
-  - sh .ci/install.sh
-
-test_script:
-  - sh .ci/test.sh
-
-on_success:
-  - sh .ci/after_success.sh
