+++ .travis.yml
@@ -1,28 +1,0 @@
-branches:
-  only:
-    - "master"
-
-os: linux
-dist: bionic
-
-language: python
-python:
-  - 3.6
-  - 3.7
-  - 3.8
-
-jobs:
-  fast_finish: true
-  include:
-    - os: osx
-      language: shell
-
-before_install: chmod +x .ci/install.sh
-install: ./.ci/install.sh
-
-before_script:
-  - chmod +x .ci/test.sh
-  - chmod +x .ci/after_success.sh
-script: ./.ci/test.sh
-
-after_success: ./.ci/after_success.sh
