+++ .github/workflows/docs.yml
@@ -31,17 +31,13 @@ jobs:
       - name: Checkout repository
         uses: actions/checkout@v2
 
-      - name: Install torch for CPU
-        run: |
-          REQUIREMENTS_FILE=github_actions_torch_requirements.txt
-          python gen_torch_cpu_requirements.py --file $REQUIREMENTS_FILE
-          pip install -r $REQUIREMENTS_FILE
-
       - name: Install doc requirements
         run: pip install -r requirements-doc.txt
 
       - name: Install pystiche
-        run: pip install .
+        run: |
+          pip install light-the-torch
+          ltt --full-install .
 
       - name: Build html
         working-directory: docs
