### 2020-07-13 3a6547e
use tox for CI tasks
