+++ .github/workflows/lint.yml
@@ -26,7 +26,7 @@ jobs:
           python-version: "3.6"
 
       - name: Upgrade and install additional system packages
-        run: pip install --upgrade pip setuptools wheel
+        run: pip install --upgrade pip setuptools virtualenv wheel
 
       - name: Checkout repository
         uses: actions/checkout@v2
