+++ .github/workflows/docs.yml
@@ -30,7 +30,7 @@ jobs:
           sudo apt install -y latexmk texlive-latex-recommended texlive-latex-extra texlive-fonts-recommended
 
       - name: Upgrade and install additional system packages
-        run: pip install --upgrade pip setuptools wheel
+        run: pip install --upgrade pip setuptools virtualenv wheel
 
       - name: Install dev requirements
         run: pip install -r requirements-dev.txt
