+++ .github/workflows/docs.yml
@@ -32,8 +32,8 @@ jobs:
       - name: Upgrade and install additional system packages
         run: pip install --upgrade pip setuptools virtualenv wheel
 
-      - name: ls
-        run: ls -lA
+      - name: Checkout repository
+        uses: actions/checkout@v2
 
       - name: Install dev requirements
         run: pip install -r requirements-dev.txt
