### 2020-07-13 d5e1c50
use tox for CI tasks (#339)
* use tox for CI tasks

* adapt workflows

* ignore tox envs during lint

* add more system packages

* debug docs failure

* fix docs workflow
