+++ tox.ini
@@ -1,0 +1,68 @@
+[tox]
+;See link below for available options
+;https://tox.readthedocs.io/en/latest/config.html
+
+requires =
+  tox-ltt
+
+envlist = py{36, 37, 38}
+skip_missing_interpreters = True
+
+[testenv]
+deps =
+  pytest
+  pyimagetest
+  pillow_affine
+  dill
+  pytest-subtests
+  coverage
+commands =
+  coverage run \
+    --rcfile=.coveragerc \
+    -m pytest \
+      -c pytest.ini \
+      {posargs}
+
+[testenv:lint]
+force_cpu = True
+deps =
+  isort < 5
+  black
+  flake8 >= 3.8
+  mypy
+  git+https://github.com/numpy/numpy-stubs
+commands =
+  isort \
+    --settings-path=pyproject.toml \
+    --check-only \
+    --recursive .
+  black \
+    --config=pyproject.toml \
+    --check \
+    .
+  flake8 \
+    --config=.flake8 \
+    .
+  mypy \
+    --config-file=mypy.ini
+
+[testenv:docs]
+passenv =
+  READTHEDOCS
+  GITHUB_ACTIONS
+  CI
+  PYSTICHE_PLOT_GALLERY
+  PYSTICHE_DOWNLOAD_GALLERY
+deps =
+  sphinx >= 2.1, < 3
+  sphinxcontrib-bibtex
+  sphinx_autodoc_typehints
+  sphinx-gallery>=0.7.0
+  # Additional sphinx-gallery dependencies
+  # https://sphinx-gallery.github.io/stable/index.html#install-via-pip
+  matplotlib
+  sphinx_rtd_theme
+changedir = docs
+commands =
+  sphinx-build -M html source build
+  sphinx-build -M latexpdf source build
