### 2020-07-13 6f8491e
use pre-commit as linter
