+++ tox.ini
@@ -24,27 +24,17 @@ commands =
       {posargs}
 
 [testenv:lint]
+whitelist_externals =
+  pre-commit
 force_cpu = True
 deps =
-  isort < 5
-  black
   flake8 >= 3.8
   mypy
   git+https://github.com/numpy/numpy-stubs
 commands =
-  isort \
-    --settings-path=pyproject.toml \
-    --check-only \
-    --recursive .
-  black \
-    --config=pyproject.toml \
-    --check \
-    .
-  flake8 \
-    --config=.flake8 \
-    .
-  mypy \
-    --config-file=mypy.ini
+  pre-commit run --all-files
+  flake8 --config=.flake8 .
+  mypy --config-file=mypy.ini
 
 [testenv:docs]
 passenv =
