+++ tox.ini
@@ -23,6 +23,15 @@ commands =
       -c pytest.ini \
       {posargs}
 
+[testenv:format]
+whitelist_externals =
+  pre-commit
+skip_install = True
+disable_light_the_torch = True
+deps =
+commands =
+  pre-commit run --all-files
+
 [testenv:lint]
 whitelist_externals =
   pre-commit
