### 2020-07-13 61be04a
use pre-commit as linter (#340)
* use pre-commit as linter

* add a format env for tox

* lint
