+++ setup.cfg
@@ -1,0 +1,47 @@
+[metadata]
+name = pystiche
+platforms = any
+description = Framework for Neural Style Transfer built upon PyTorch
+long_description = file: README.rst
+long_description_content_type = text/x-rst
+keywords = neural-style-transfer, framework, pytorch
+url = https://github.com/pmeier/pystiche
+author = Philip Meier
+author-email = github.pmeier@posteo.de
+license = BSD-3-Clause
+classifiers =
+    Development Status :: 4 - Beta
+    Intended Audience :: Developers
+    Intended Audience :: Education
+    Intended Audience :: Science/Research
+    License :: OSI Approved :: BSD License
+    Programming Language :: Python :: 3.6
+    Programming Language :: Python :: 3.7
+    Programming Language :: Python :: 3.8
+    Topic :: Scientific/Engineering
+    Topic :: Scientific/Engineering :: Artificial Intelligence
+    Topic :: Software Development :: Libraries :: Application Frameworks
+    Topic :: Software Development :: Libraries :: Python Modules
+    Typing :: Typed
+project_urls =
+    Bug Tracker = https://github.com/pmeier/pystiche/issues
+    Documentation = https://pystiche.readthedocs.io
+    Source Code = https://github.com/pmeier/pystiche
+
+
+[options]
+packages = find:
+include_package_data = True
+python_requires = >=3.6
+install_requires =
+    torch>=1.5.0
+    torchvision>=0.6.0
+    pillow
+    numpy
+    requests
+    typing_extensions
+
+[options.packages.find]
+exclude =
+    tests
+    tests.*
