+++ tox.ini
@@ -5,6 +5,7 @@
 requires =
   tox-ltt
 
+isolated_build = True
 envlist = py{36, 37, 38}
 skip_missing_interpreters = True
 
