+++ pystiche/_version.py
@@ -1,0 +1,4 @@
+# coding: utf-8
+# file generated by setuptools_scm
+# don't change, don't track in version control
+version = "0.5.1.dev9+g05aeb3e.d20200713145926"
