+++ pyproject.toml
@@ -10,7 +10,7 @@ build-backend = "setuptools.build_meta"
 # See link below for available options
 # https://github.com/pypa/setuptools_scm/#configuration-parameters
 
-write_to = "pystiche_papers/_version.py"
+write_to = "pystiche/_version.py"
 version_scheme = "guess-next-dev"
 local_scheme = "node-and-timestamp"
 
