### 2020-07-13 3f69c52
fix smoke tests
