+++ mypy.ini
@@ -24,6 +24,9 @@ show_error_context = True
 show_error_codes = True
 pretty = True
 
+[mypy-pystiche]
+warn_unused_ignores = False
+
 [mypy-torchvision.*]
 ignore_missing_imports = True
 
