### 2020-07-13 8ea419e
add workflow to test if package is publishable
