+++ tox.ini
@@ -67,3 +67,18 @@ changedir = docs
 commands =
   sphinx-build -M html source build
   sphinx-build -M latexpdf source build
+
+[testenv:publishable]
+whitelist_externals =
+  rm
+skip_install = True
+disable_light_the_torch = True
+deps =
+  check-wheel-contents
+  pep517
+  twine
+commands =
+  rm -rf build dist pystiche_papers.egg-info
+  python -m pep517.build --source --binary .
+  twine check dist/*
+  check-wheel-contents dist
