### 2020-07-13 3493bb2
enable all workflow on release branches
