+++ .github/workflows/lint.yml
@@ -4,6 +4,7 @@ on:
   push:
     branches:
       - master
+      - releases/*
 
   pull_request:
     paths:
