+++ tox.ini
@@ -55,6 +55,7 @@ passenv =
   PYSTICHE_DOWNLOAD_GALLERY
 deps =
   sphinx >= 2.1, < 3
+  importlib-metadata
   sphinxcontrib-bibtex
   sphinx_autodoc_typehints
   sphinx-gallery>=0.7.0
