### 2020-07-13 2c4534f
fix rtd configs
