### 2020-07-13 41b67c6
update RTD requirements
