+++ .github/workflows/lint.yml
@@ -32,6 +32,8 @@ jobs:
 
       - name: Checkout repository
         uses: actions/checkout@v2
+        with:
+          fetch-depth: 0
 
       - name: Install dev requirements
         run: pip install -r requirements-dev.txt
