+++ .github/workflows/tests.yml
@@ -42,6 +42,8 @@ jobs:
 
       - name: Checkout repository
         uses: actions/checkout@v2
+        with:
+          fetch-depth: 0
 
       - name: Install dev requirements
         run: pip install -r requirements-dev.txt
