+++ .github/workflows/docs.yml
@@ -36,6 +36,8 @@ jobs:
 
       - name: Checkout repository
         uses: actions/checkout@v2
+        with:
+          fetch-depth: 0
 
       - name: Install dev requirements
         run: pip install -r requirements-dev.txt
