### 2020-07-14 ae14391
fix version in docs conf
