### 2020-07-14 e2e5136
embrace PEP517 (#341)
* embrace PEP517

* update workflow triggers

* fix docs

* fix version

* fix smoke tests

* cleanup

* fix docs

* fix

* fix download

* lint

* fix download

* fix rtd configs

* update RTD requirements

* increase git checkout fetch depth

* fix version in docs conf
