+++ pystiche/__about__.py
@@ -1,21 +1,0 @@
-__all__ = [
-    "__name__",
-    "__description__",
-    "__keywords__",
-    "__base_version__",
-    "__is_dev_version__",
-    "__url__",
-    "__license__",
-    "__author__",
-    "__author_email__",
-]
-
-__name__ = "pystiche"
-__description__ = "Framework for Neural Style Transfer (NST) built upon PyTorch"
-__keywords__ = "neural-style-transfer framework pytorch"
-__base_version__ = "1.0.0"
-__is_dev_version__ = True
-__url__ = "https://github.com/pmeier/pystiche"
-__license__ = "BSD 3-Clause"
-__author__ = "Philip Meier"
-__author_email__ = "github.pmeier@posteo.de"
