+++ tox.ini
@@ -5,6 +5,7 @@
 requires =
   tox-ltt
 
+isolated_build = True
 envlist = py{36, 37, 38}
 skip_missing_interpreters = True
 
@@ -54,6 +55,7 @@ passenv =
   PYSTICHE_DOWNLOAD_GALLERY
 deps =
   sphinx >= 2.1, < 3
+  importlib-metadata
   sphinxcontrib-bibtex
   sphinx_autodoc_typehints
   sphinx-gallery>=0.7.0
