+++ .github/workflows/lint.yml
@@ -14,6 +14,8 @@ on:
       - "tox.ini"
       - "requirements-dev.txt"
       - ".github/workflows/lint.yml"
+      - "setup.cfg"
+      - "pyproject.toml"
 
 jobs:
   check:
@@ -30,6 +32,8 @@ jobs:
 
       - name: Checkout repository
         uses: actions/checkout@v2
+        with:
+          fetch-depth: 0
 
       - name: Install dev requirements
         run: pip install -r requirements-dev.txt
