### 2020-07-14 7689849
ad DOI badge to README (#343)
