### 2020-07-14 5f5f5ef
add workflow to test if package is publishable (#342)
* add workflow to test if package is publishable

* enable all workflow on release branches

* add manifest

* fix workflow

* rename workflow steps
