+++ .github/workflows/publishable.yml
@@ -1,4 +1,4 @@
-name: lint
+name: publishable
 
 on:
   push:
@@ -8,25 +8,27 @@ on:
 
   pull_request:
     paths:
-      - "**.py"
+      - "pystiche/**"
+      - "CONTRIBUTING.rst"
+      - ".github/workflows/publishable.yml"
+      - ".gitignore"
+      - "LICENSE"
+      - "MANIFEST.in"
       - "pyproject.toml"
-      - ".flake8"
-      - "mypy.ini"
-      - "tox.ini"
       - "requirements-dev.txt"
-      - ".github/workflows/lint.yml"
+      - "README.rst"
       - "setup.cfg"
-      - "pyproject.toml"
+      - "tox.ini"
 
 jobs:
-  check:
+  pypi:
     runs-on: ubuntu-latest
 
     steps:
       - name: Set up python
         uses: actions/setup-python@v2
         with:
-          python-version: "3.6"
+          python-version: 3.6
 
       - name: Upgrade and install additional system packages
         run: pip install --upgrade pip setuptools virtualenv wheel
@@ -40,7 +42,7 @@ jobs:
         run: pip install -r requirements-dev.txt
 
       - name: Create environment
-        run: tox -e lint --notest
+        run: tox -e publishable --notest
 
-      - name: Run lint
-        run: tox -e lint
+      - name: Test if publishable
+        run: tox -e publishable
