### 2020-07-15 #25 @NickleDave
Hi @pmeier thank you for submitting and finding reviewers!
And thank you for your patience -- I should have replied sooner.

@edgarriba and @soumith welcome!
It seems like Philip has definitely brought people with the right expertise to review the package.
