### 2020-07-15 #25 @NickleDave
@lwasser here are my editor checks:

## Editor checks:

- [x] **License:** The package has an OSI accepted license
