### 2020-07-15 #25 @NickleDave
## Editor comments

@lwasser I will follow up this weekend with editor comments

---

Reviewers: @edgarriba @soumith
Due date:

@edgarriba @soumith typically we ask for 2-3 week turnaround for reviews.
Although we should see if we can get editors to meet that standard first (I'm sorry @pmeier -- thank you again for your patience!). And I know everything is extra crazy for everyone right now.

Can you both let me know how doable three weeks sounds to you?

* Our guidelines for reviewers are here: https://www.pyopensci.org/contributing-guide/open-source-software-submissions/reviewer-guide.html
* There's a template for reviews here:
https://www.pyopensci.org/contributing-guide/appendices/templates.html
* And here's an example of reviewers filling out the template are on these previous reviews:
https://github.com/pyOpenSci/software-review/issues/12#issuecomment-527622205
