### 2020-07-15 #25 @NickleDave
Excellent, thank you @edgarriba. Makes sense
Setting the due date for Aug 5 -- cc @soumith @pmeier @lwasser
