### 2020-07-15 #25 @pmeier
@NickleDave

Note that we probably need a third reviewer since, to my knowledge, both, @edgarriba and @soumith, have no domain knowledge about Neural Style Transfer (NST). Shall I assist the search?
