### 2020-07-16 a0568f9
add auto generation for RTD requirements
