+++ setup.cfg
@@ -33,6 +33,8 @@ project_urls =
 packages = find:
 include_package_data = True
 python_requires = >=3.6
+# Remember to re-run docs/generate_requirements_rtd.py if you change the install
+# requirements
 install_requires =
     torch>=1.5.0
     torchvision>=0.6.0
