### 2020-07-16 a22ea9c
add auto generation for RTD requirements (#345)
* add auto generation for RTD requirements

* add reminders
