+++ tox.ini
@@ -53,6 +53,7 @@ passenv =
   CI
   PYSTICHE_PLOT_GALLERY
   PYSTICHE_DOWNLOAD_GALLERY
+# Remember to re-run docs/generate_requirements_rtd.py if you change the dependencies
 deps =
   sphinx >= 2.1, < 3
   importlib-metadata
