### 2020-07-20 #25 @pmeier
@edgarriba, @soumith

I just saw that AppVeyor Badge for `v0.5.0` is not working properly:

This is due to the fact that I switched to GitHub Actions for the current `master`, and AppVeyor kept failing since `pystiche` was still activated. After I deactivated it, the badge went to `Project unknown`, which was not better either. I tried to re-run for `v0.5.0`, but AFAIK that is not possible. Re-running from `master` always fails, since the needed scripts no longer exist. Thus, the badge is now stuck at `cancelled`.

I made no changes to the functionality or the test-suite of `pystiche` since `v0.5.0`. You can observe the results [here](https://github.com/pmeier/pystiche/actions?query=workflow%3Atests+branch%3Amaster).
