### 2020-07-21 #25 @NickleDave
@pmeier @lwasser  just want to follow up about reviewers and editor comments

For me, the main fit with pyopensci for this package is in terms of reproducibility + education.
You aim to provide a package that encapsulates key concepts from this area of research, making them easier to learn about, and making it possible to reproduce key results without boilerplate.

To that end, the ideal reviewer would be able to speak to whether you are achieving those goals.

I also think **we** should find at least one of the reviewers (and when I say we, that means: me) to make sure the process is fair. I'm sure you agree. I put out an ask on Twitter (thank you Leah for the retweet) and I have at least one person in mind that I am contacting now. Will update as soon as I hear back
