### 2020-07-21 #25 @pmeier
@NickleDave

> For me, the main fit with pyopensci for this package is in terms of reproducibility + education.
You aim to provide a package that encapsulates key concepts from this area of research, making them easier to learn about, and making it possible to reproduce key results without boilerplate.

1. Its not only **re**-producibility, but also "producibility". While I certainly aim to reproduce already published papers, my hope is that researchers use `pystiche` in upcoming publications. Reproducibility was just the most fitting out the available categories.
2. I'm not sure of the _education_ part. It is certainly a goal for the future, albeit a secondary one. Unless you mean easing the on-boarding in the field, `pystiche` currently targets other researchers.

> I also think **we** should find at least one of the reviewers (and when I say we, that means: me) to make sure the process is fair. I'm sure you agree.

Agreed. I haven't heard back from anyone yet anyway (not that I'm hyper connected since this will be my first major contribution to the field). If someone does get back to me, I'll fill them in about this and refer them to you.
