### 2020-07-23 2d4987a
add image / guide getter for OperatorContainer
