### 2020-07-23 e8a98e7
adapt structure for setters
