### 2020-07-23 b0a4152
add getter to API doc
