### 2020-07-23 ab71e6f
add image / guide getter to OperatorContainer (#347)
* add image / guide getter for OperatorContainer

* add tests

* adapt structure for setters

* typo

* add getter to API doc

* lint
