### 2020-07-25 #25 @NickleDave
Thank you @pmeier I hear you, and I think we are on the same page.
My comments are specifically about fit with PyOpenSci criteria but I don't mean to limit the scope of your package.
JOSS reviewers for example can focus on how it enables new research.
And by education I do mean "easing the on-boarding in the field". To me this is a central goal for scientific packages: providing abstractions that make the research easier to understand. The package should enable "coding at the speed of thought". Which lines up with your point 1 above.

I am still reaching out to reviewers.
@edgarriba @soumith as soon as a third reviewer joins us I will set the review due date as three weeks from that day
