+++ setup.cfg
@@ -36,8 +36,8 @@ python_requires = >=3.6
 # Remember to re-run docs/generate_requirements_rtd.py if you change the install
 # requirements
 install_requires =
-    torch>=1.5.0
-    torchvision>=0.6.0
+    torch>=1.6.0
+    torchvision>=0.7.0
     pillow
     numpy
     requests
