### 2020-07-29 a9b0673
invoke pip with python when self upgrading
