+++ .github/workflows/docs.yml
@@ -32,7 +32,10 @@ jobs:
           sudo apt update -y
           sudo apt install -y latexmk texlive-latex-recommended texlive-latex-extra texlive-fonts-recommended
 
-      - name: Upgrade and install additional system packages
+      - name: Upgrade pip
+        run: python -m pip install --upgrade pip
+
+      - name: Upgrade or install additional system packages
         run: pip install --upgrade pip setuptools virtualenv wheel
 
       - name: Checkout repository
