+++ .github/workflows/publishable.yml
@@ -30,7 +30,10 @@ jobs:
         with:
           python-version: 3.6
 
-      - name: Upgrade and install additional system packages
+      - name: Upgrade pip
+        run: python -m pip install --upgrade pip
+
+      - name: Upgrade or install additional system packages
         run: pip install --upgrade pip setuptools virtualenv wheel
 
       - name: Checkout repository
