### 2020-07-29 e421353
invoke pip with python when self upgrading (#349)
