+++ .github/workflows/tests.yml
@@ -38,8 +38,11 @@ jobs:
         with:
           python-version: ${{ matrix.python }}
 
-      - name: Upgrade and install additional system packages
-        run: pip install --upgrade pip virtualenv setuptools wheel
+      - name: Upgrade pip
+        run: python -m pip install --upgrade pip
+
+      - name: Upgrade or install additional system packages
+        run: pip install --upgrade pip setuptools virtualenv wheel
 
       - name: Checkout repository
         uses: actions/checkout@v2
