+++ setup.py
@@ -65,8 +65,8 @@ package_data = {PACKAGE_NAME: [version_file, "py.typed"]}
 
 
 install_requires = (
-    "torch>=1.5.0",
-    "torchvision>=0.6.0",
+    "torch >=1.5, <1.6",
+    "torchvision >=0.6, <0.7",
     "pillow",
     "numpy",
     "requests",
