+++ pystiche/__about__.py
@@ -13,7 +13,7 @@ __all__ = [
 __name__ = "pystiche"
 __description__ = "Framework for Neural Style Transfer (NST) built upon PyTorch"
 __keywords__ = "neural-style-transfer framework pytorch"
-__base_version__ = "0.5.0"
+__base_version__ = "0.5.0.post0"
 __is_dev_version__ = False
 __url__ = "https://github.com/pmeier/pystiche"
 __license__ = "BSD 3-Clause"
