### 2020-07-29 #25 @pmeier
@edgarriba, @soumith, @NickleDave

I've just released [`pystiche==0.5.0.post0`](https://github.com/pmeier/pystiche/releases/tag/v0.5.0.post0) since yesterdays release of `torch==1.6.0` broke compatibility.
