+++ tox.ini
@@ -11,7 +11,7 @@ skip_missing_interpreters = True
 
 [testenv]
 deps =
-  pytest
+  pytest >= 6
   pyimagetest
   pillow_affine
   dill
