### 2020-07-29 25ef80e
Bump PyTorch versions (#348)
* bump PyTorch versions

* lint

* temporary test fix for compatibility

* add comment referencing PR
