### 2020-07-29 0f100a0
try remove win compatibility code
