### 2020-07-29 d8762a1
Revert "try remove win compatibility code"
This reverts commit 0f100a06ff66451db87113c9665c0504a4c65691.
