### 2020-07-29 f1c507e
try closing only used file handler
