### 2020-07-29 dd65292
pytest>=6 compatibility (#350)
* pytest>=6 compatibility

* try remove win compatibility code

* Revert "try remove win compatibility code"

This reverts commit 0f100a06ff66451db87113c9665c0504a4c65691.

* try closing only used file handler
