### 2020-07-29 4a06ac8
update RTD requirements (#351)
