### 2020-07-30 2a4b018
upgrade sphinx and sphinx-autodoc-typehints
