### 2020-07-30 460a684
pass HOME env var for latex build
