+++ tox.ini
@@ -48,6 +48,7 @@ commands =
 
 [testenv:docs]
 passenv =
+  HOME
   READTHEDOCS
   GITHUB_ACTIONS
   CI
