### 2020-07-30 a128f1e
upgrade RTD requirements
