+++ docs/requirements-rtd.txt
@@ -1,9 +1,9 @@
-sphinx >= 2.1, < 3
+sphinx >= 3
 importlib-metadata
 sphinxcontrib-bibtex
-sphinx_autodoc_typehints
+sphinx_autodoc_typehints >= 1.11
 sphinx-gallery>=0.7.0
 matplotlib
 sphinx_rtd_theme
-https://download.pytorch.org/whl/cpu/torchvision-0.7.0%2Bcpu-cp36-cp36m-linux_x86_64.whl
 https://download.pytorch.org/whl/cpu/torch-1.6.0%2Bcpu-cp36-cp36m-linux_x86_64.whl
+https://download.pytorch.org/whl/cpu/torchvision-0.7.0%2Bcpu-cp36-cp36m-linux_x86_64.whl
