+++ .github/workflows/docs.yml
@@ -22,10 +22,10 @@ jobs:
     runs-on: ubuntu-latest
 
     steps:
-      - name: Set up python
-        uses: actions/setup-python@v2
-        with:
-          python-version: "3.6"
+#      - name: Set up python
+#        uses: actions/setup-python@v2
+#        with:
+#          python-version: "3.6"
 
       - name: Install latex requirements
         run: |
@@ -34,6 +34,9 @@ jobs:
 
       - name: Check config
         run: kpsewhich -a texmf.cnf
+
+      - name: Display config
+        run: nl /usr/share/texlive/texmf-dist/web2c/texmf.cnf
 
 #      - name: Upgrade pip
 #        run: python -m pip install --upgrade pip
