+++ .github/workflows/docs.yml
@@ -22,10 +22,10 @@ jobs:
     runs-on: ubuntu-latest
 
     steps:
-#      - name: Set up python
-#        uses: actions/setup-python@v2
-#        with:
-#          python-version: "3.6"
+      - name: Set up python
+        uses: actions/setup-python@v2
+        with:
+          python-version: "3.6"
 
       - name: Install latex requirements
         run: |
@@ -37,6 +37,11 @@ jobs:
 
       - name: Display config
         run: nl /usr/share/texlive/texmf-dist/web2c/texmf.cnf
+
+      - name: Increase memory
+        run: |
+          sudo echo "main_memory = 500000000" > /usr/share/texlive/texmf.cnf
+          nl /usr/share/texlive/texmf.cnf
 
 #      - name: Upgrade pip
 #        run: python -m pip install --upgrade pip
