### 2020-07-30 45dfab0
increase memory in global config
