+++ .github/workflows/docs.yml
@@ -43,26 +43,26 @@ jobs:
           sudo sed -i "s/main_memory = \([0-9]\+\).*/main_memory = \100/" /usr/share/texlive/texmf-dist/web2c/texmf.cnf
           nl /usr/share/texlive/texmf-dist/web2c/texmf.cnf | grep main_memory
 
-#      - name: Upgrade pip
-#        run: python -m pip install --upgrade pip
-#
-#      - name: Upgrade or install additional system packages
-#        run: pip install --upgrade pip setuptools virtualenv wheel
-#
-#      - name: Checkout repository
-#        uses: actions/checkout@v2
-#        with:
-#          fetch-depth: 0
-#
-#      - name: Install dev requirements
-#        run: pip install -r requirements-dev.txt
-#
-#      - name: Create environment
-#        run: tox -e docs --notest
-#
-#      - name: Build documentation
-#        run: tox -e docs
-#
+      - name: Upgrade pip
+        run: python -m pip install --upgrade pip
+
+      - name: Upgrade or install additional system packages
+        run: pip install --upgrade pip setuptools virtualenv wheel
+
+      - name: Checkout repository
+        uses: actions/checkout@v2
+        with:
+          fetch-depth: 0
+
+      - name: Install dev requirements
+        run: pip install -r requirements-dev.txt
+
+      - name: Create environment
+        run: tox -e docs --notest
+
+      - name: Build documentation
+        run: tox -e docs
+
 #      - name: Upload html build
 #        uses: actions/upload-artifact@v2
 #        with:
