### 2020-07-30 57397fb
reduce memory increase by factor 10
