+++ tox.ini
@@ -66,7 +66,7 @@ deps =
   sphinx_rtd_theme
 changedir = docs
 commands =
-  sphinx-build -M html source build
+;  sphinx-build -M html source build
   sphinx-build -M latexpdf source build
 
 [testenv:publishable]
