+++ .github/workflows/docs.yml
@@ -39,9 +39,7 @@ jobs:
         run: nl /usr/share/texlive/texmf-dist/web2c/texmf.cnf
 
       - name: Increase memory
-        run: |
-          sudo sed -i "s/main_memory = \([0-9]\+\).*/main_memory = \10/" /usr/share/texlive/texmf-dist/web2c/texmf.cnf
-          nl /usr/share/texlive/texmf-dist/web2c/texmf.cnf | grep main_memory
+        run: sudo sed -i "s/main_memory = .*/main_memory = 256000000/" /usr/share/texlive/texmf-dist/web2c/texmf.cnf
 
       - name: Upgrade pip
         run: python -m pip install --upgrade pip
