+++ .github/workflows/publishable.yml
@@ -10,7 +10,7 @@ on:
     paths:
       - "pystiche/**"
       - "CONTRIBUTING.rst"
-      - ".github/workflows/publishable.yml"
+#      - ".github/workflows/publishable.yml"
       - ".gitignore"
       - "LICENSE"
       - "MANIFEST.in"
