### 2020-07-30 48784e5
increase memory in every config
