+++ .github/workflows/docs.yml
@@ -43,6 +43,7 @@ jobs:
           sudo sed -i "s/main_memory = .*/main_memory = 12435455/" /etc/texmf/web2c/texmf.cnf
           sudo sed -i "s/main_memory = .*/main_memory = 12435455/" /usr/share/texmf/web2c/texmf.cnf
           sudo sed -i "s/main_memory = .*/main_memory = 12435455/" /usr/share/texlive/texmf-dist/web2c/texmf.cnf
+          echo "12435455" | sudo tee /usr/share/texlive/texmf.cnf
 
       - name: Upgrade pip
         run: python -m pip install --upgrade pip
