### 2020-07-30 5916e93
temporarily disable latex doc builds
