+++ tox.ini
@@ -67,7 +67,7 @@ deps =
 changedir = docs
 commands =
   sphinx-build -M html source build
-  sphinx-build -M latexpdf source build
+;  sphinx-build -M latexpdf source build
 
 [testenv:publishable]
 whitelist_externals =
