### 2020-07-30 3003901
temporarily disable latex doc builds (#355)
* temporarily disable latex doc builds

* disable latex artifact upload
