### 2020-07-30 8c28cce
Revert "temporarily disable latex doc builds (#355)"
This reverts commit 30039013fb452a5abed7dd4ff637c06d0ae98960.
