+++ .readthedocs.yml
@@ -1,14 +1,9 @@
-#See link below for available options
-#https://docs.readthedocs.io/en/stable/config-file/
-
 version: 2
 
 sphinx:
   configuration: docs/source/conf.py
 
-formats:
-  - htmlzip
-  - epub
+formats: all
 
 python:
   version: 3.6
