### 2020-07-30 0920e30
Bump sphinx to >= 3 (#353)
* upgrade sphinx and sphinx-autodoc-typehints

* pass HOME env var for latex build

* upgrade RTD requirements
