+++ tox.ini
@@ -48,6 +48,7 @@ commands =
 
 [testenv:docs]
 passenv =
+  HOME
   READTHEDOCS
   GITHUB_ACTIONS
   CI
@@ -55,10 +56,10 @@ passenv =
   PYSTICHE_DOWNLOAD_GALLERY
 # Remember to re-run docs/generate_requirements_rtd.py if you change the dependencies
 deps =
-  sphinx >= 2.1, < 3
+  sphinx >= 3
   importlib-metadata
   sphinxcontrib-bibtex
-  sphinx_autodoc_typehints
+  sphinx_autodoc_typehints >= 1.11
   sphinx-gallery>=0.7.0
   # Additional sphinx-gallery dependencies
   # https://sphinx-gallery.github.io/stable/index.html#install-via-pip
