### 2020-07-30 4eadf70
split tox test envs for granular control
