+++ tox.ini
@@ -3,13 +3,12 @@
 ;https://tox.readthedocs.io/en/latest/config.html
 
 requires =
+  tox-factor
   tox-ltt
-
+envlist = tests
 isolated_build = True
-envlist = py{36, 37, 38}
-skip_missing_interpreters = True
 
-[testenv]
+[testenv:tests]
 deps =
   pytest >= 6
   pyimagetest
@@ -24,29 +23,29 @@ commands =
       -c pytest.ini \
       {posargs}
 
-[testenv:format]
+[testenv:lint-format]
 whitelist_externals =
   pre-commit
 skip_install = True
-disable_light_the_torch = True
-deps =
 commands =
   pre-commit run --all-files
 
-[testenv:lint]
-whitelist_externals =
-  pre-commit
-force_cpu = True
+[testenv:lint-style]
 deps =
   flake8 >= 3.8
+skip_install = True
+commands =
+  flake8 --config=.flake8 .
+
+[testenv:lint-typing]
+force_cpu = True
+deps =
   mypy
   git+https://github.com/numpy/numpy-stubs
 commands =
-  pre-commit run --all-files
-  flake8 --config=.flake8 .
   mypy --config-file=mypy.ini
 
-[testenv:docs]
+[docs-common]
 passenv =
   HOME
   READTHEDOCS
@@ -66,9 +65,20 @@ deps =
   matplotlib
   sphinx_rtd_theme
 changedir = docs
+
+[testenv:docs-html]
+passenv = {[docs-common]passenv}
+deps = {[docs-common]deps}
+changedir = {[docs-common]changedir}
 commands =
   sphinx-build -M html source build
-;  sphinx-build -M latexpdf source build
+
+[testenv:docs-pdf]
+passenv = {[docs-common]passenv}
+deps = {[docs-common]deps}
+changedir = {[docs-common]changedir}
+commands =
+  sphinx-build -M latexpdf source build
 
 [testenv:publishable]
 whitelist_externals =
