+++ tox.ini
@@ -23,19 +23,12 @@ commands =
       -c pytest.ini \
       {posargs}
 
-[testenv:lint-format]
+[testenv:lint-style]
 whitelist_externals =
   pre-commit
 skip_install = True
 commands =
   pre-commit run --all-files
-
-[testenv:lint-style]
-deps =
-  flake8 >= 3.8
-skip_install = True
-commands =
-  flake8 --config=.flake8 .
 
 [testenv:lint-typing]
 force_cpu = True
