+++ .pre-commit-config.yaml
@@ -10,6 +10,11 @@ repos:
     hooks:
       - id: black
         args: [--config=pyproject.toml]
+  - repo: https://gitlab.com/pycqa/flake8
+    rev: "3.8.3"
+    hooks:
+      - id: flake8
+        args: [--config=.flake8]
   - repo: https://github.com/pre-commit/pre-commit-hooks
     rev: v3.1.0
     hooks:
