+++ tox.ini
@@ -66,7 +66,7 @@ changedir = {[docs-common]changedir}
 commands =
   sphinx-build -M html source build
 
-[testenv:docs-pdf]
+[testenv:docs-latex]
 passenv = {[docs-common]passenv}
 deps = {[docs-common]deps}
 changedir = {[docs-common]changedir}
