+++ .github/workflows/tests.yml
@@ -58,6 +58,9 @@ jobs:
       - name: Run tests
         run: tox -- --skip-large-download
 
+      - name: List dir
+        run: ls .
+
       - name: Report coverage
         run: |
           pip install coverage
