+++ .github/workflows/tests.yml
@@ -58,14 +58,6 @@ jobs:
       - name: Run tests
         run: tox -- --skip-large-download
 
-      - name: List dir
-        run: ls .
-
-      - name: Report coverage
-        run: |
-          pip install coverage
-          coverage report
-
       - name: Upload coverage
         uses: codecov/codecov-action@v1.0.12
         with:
