### 2020-07-30 cda4699
split tox testenvs for more granular control (#356)
* split tox test envs for granular control

* add flake8 to pre-commit

* adapt workflows

* update RTD requirements generation

* add coverage report

* add debug options

* bump codecov action version

* cleanup
