+++ .github/workflows/docs.yml
@@ -18,7 +18,7 @@ on:
       - "pyproject.toml"
 
 jobs:
-  build:
+  html:
     runs-on: ubuntu-latest
 
     steps:
@@ -26,11 +26,6 @@ jobs:
         uses: actions/setup-python@v2
         with:
           python-version: "3.6"
-
-      - name: Install latex requirements
-        run: |
-          sudo apt update -y
-          sudo apt install -y latexmk texlive-latex-recommended texlive-latex-extra texlive-fonts-recommended
 
       - name: Upgrade pip
         run: python -m pip install --upgrade pip
@@ -47,18 +42,52 @@ jobs:
         run: pip install -r requirements-dev.txt
 
       - name: Create environment
-        run: tox -e docs --notest
+        run: tox -e docs-html --notest
 
-      - name: Build documentation
-        run: tox -e docs
+      - name: Build html documentation
+        run: tox -e docs-html
 
-      - name: Upload html build
+      - name: Upload build
         uses: actions/upload-artifact@v2
         with:
           name: html
           path: docs/build/html
 
-#      - name: Upload latex build
+#  latex:
+#    runs-on: ubuntu-latest
+#
+#    steps:
+#      - name: Set up python
+#        uses: actions/setup-python@v2
+#        with:
+#          python-version: "3.6"
+#
+#      - name: Install latex requirements
+#        run: |
+#          sudo apt update -y
+#          sudo apt install -y latexmk texlive-latex-recommended texlive-latex-extra texlive-fonts-recommended
+#
+#      - name: Upgrade pip
+#        run: python -m pip install --upgrade pip
+#
+#      - name: Upgrade or install additional system packages
+#        run: pip install --upgrade pip setuptools virtualenv wheel
+#
+#      - name: Checkout repository
+#        uses: actions/checkout@v2
+#        with:
+#          fetch-depth: 0
+#
+#      - name: Install dev requirements
+#        run: pip install -r requirements-dev.txt
+#
+#      - name: Create environment
+#        run: tox -e docs-latex --notest
+#
+#      - name: Build latex documentation
+#        run: tox -e docs-latex
+#
+#      - name: Upload build
 #        uses: actions/upload-artifact@v2
 #        with:
 #          name: latex
