+++ tox.ini
@@ -3,13 +3,12 @@
 ;https://tox.readthedocs.io/en/latest/config.html
 
 requires =
+  tox-factor
   tox-ltt
-
+envlist = tests
 isolated_build = True
-envlist = py{36, 37, 38}
-skip_missing_interpreters = True
 
-[testenv]
+[testenv:tests]
 deps =
   pytest >= 6
   pyimagetest
@@ -24,29 +23,22 @@ commands =
       -c pytest.ini \
       {posargs}
 
-[testenv:format]
+[testenv:lint-style]
 whitelist_externals =
   pre-commit
 skip_install = True
-disable_light_the_torch = True
-deps =
 commands =
   pre-commit run --all-files
 
-[testenv:lint]
-whitelist_externals =
-  pre-commit
+[testenv:lint-typing]
 force_cpu = True
 deps =
-  flake8 >= 3.8
   mypy
   git+https://github.com/numpy/numpy-stubs
 commands =
-  pre-commit run --all-files
-  flake8 --config=.flake8 .
   mypy --config-file=mypy.ini
 
-[testenv:docs]
+[docs-common]
 passenv =
   HOME
   READTHEDOCS
@@ -66,9 +58,20 @@ deps =
   matplotlib
   sphinx_rtd_theme
 changedir = docs
+
+[testenv:docs-html]
+passenv = {[docs-common]passenv}
+deps = {[docs-common]deps}
+changedir = {[docs-common]changedir}
 commands =
   sphinx-build -M html source build
-;  sphinx-build -M latexpdf source build
+
+[testenv:docs-latex]
+passenv = {[docs-common]passenv}
+deps = {[docs-common]deps}
+changedir = {[docs-common]changedir}
+commands =
+  sphinx-build -M latexpdf source build
 
 [testenv:publishable]
 whitelist_externals =
