### 2020-08-01 367a060
Revert "disable publication after release creation (#358)"
This reverts commit f25f77ae7035dde8614c3a2ad93d387ed21b9e28.
