### 2020-08-01 007e0d6
update contributing guidelines (#360)
