+++ CONTRIBUTING.rst
@@ -8,107 +8,126 @@ discussion. If you planning to contribute new features, please open an
 `issue <https://github.com/pmeier/pystiche/issues>`_ and discuss the feature with us
 first.
 
+To start working on ``pystiche`` clone from the latest version and install the
+development requirements:
+
+.. code-block:: sh
+
+  PYSTICHE_ROOT = pystiche_papers
+  git clone https://github.com/pmeier/pystiche $PYSTICHE_ROOT
+  cd $PYSTICHE_ROOT
+  pip install -r requirements-dev.txt
+  pre-commit install
+
 Every PR is subjected to multiple checks that it has to pass before it can be merged.
-Below you can find details about these checks and instructions how to run them locally.
+The checks are performed by `tox <https://tox.readthedocs.io/en/latest/>`_ . Below
+you can find details and instructions how to run the checks locally.
 
-Lint
-----
-
-.. note::
-
-  You can install all lint requirements with
-
-  .. code-block:: sh
-
-    pip install -r $PYSTICHE_ROOT/requirements-lint.txt
+Code format and linting
+-----------------------
 
 ``pystiche`` uses `isort <https://timothycrosley.github.io/isort/>`_ to sort the
 imports, `black <https://black.readthedocs.io/en/stable/>`_ to format the code, and
 `flake8 <https://flake8.pycqa.org/en/latest/>`_ to enforce
-`PEP8 <https://www.python.org/dev/peps/pep-0008/>`_ compliance.
-
-Furthermore, ``pystiche`` is `PEP561 <https://www.python.org/dev/peps/pep-0561/>`_
-compliant and checks the type annotations with `mypy <http://mypy-lang.org/>`_ .
-
-You can run all tools locally with:
+`PEP8 <https://www.python.org/dev/peps/pep-0008/>`_ compliance. To format and check the
+code style, run
 
 .. code-block:: sh
 
   cd $PYSTICHE_ROOT
-  isort --recursive .
-  black .
-  flake8
-  mypy
+  tox -e lint-style
 
 .. note::
 
-  The checks with ``isort``, ``black``, and ``flake8`` can be executed as a pre-commit
-  hook. You can install them with:
+  Amongst others, ``isort``, ``black``, and ``flake8`` are run by
+  `pre-commit <https://pre-commit.com/>`_ before every commit.
 
-  .. code-block:: sh
-
-    pip install pre-commit
-    cd $PYSTICHE_ROOT
-    pre-commit install
-
-  ``mypy`` is excluded from this, since the pre-commit runs in a separate virtual
-  environment in which ``pystiche`` would have to be installed in for every commit.
-
-Test
-----
-
-.. note::
-
-  You can install all test requirements with
-
-  .. code-block:: sh
-
-    pip install -r $PYSTICHE_ROOT/requirements-test.txt
-
-``pystiche`` uses `pytest <https://docs.pytest.org/en/stable/>`_ to run the test suite.
-You can run it locally with:
+Furthermore, ``pystiche_papers`` is
+`PEP561 <https://www.python.org/dev/peps/pep-0561/>`_ compliant and checks the type
+annotations with `mypy <http://mypy-lang.org/>`_ . To check the static typing, run
 
 .. code-block:: sh
 
-  cd $PYSTICHE_HOME
-  pytest
+  cd $PYSTICHE_ROOT
+  tox -e lint-typing
+
+You can run all lint checks with
+
+.. code-block:: sh
+
+  cd $PYSTICHE_ROOT
+  tox -f lint
+
+
+Tests
+-----
+
+``pystiche`` uses `pytest <https://docs.pytest.org/en/stable/>`_ to run the test suite.
+You can run it locally with
+
+.. code-block:: sh
+
+  cd $PYSTICHE_ROOT
+  tox
+
+.. note::
+
[…]
