### 2020-08-07 165cd76
disable all formats other than pdf
