+++ .readthedocs.yml
@@ -3,7 +3,7 @@ version: 2
 sphinx:
   configuration: docs/source/conf.py
 
-formats: all
+formats: pdf
 
 python:
   version: 3.6
