+++ .readthedocs.yml
@@ -3,7 +3,8 @@ version: 2
 sphinx:
   configuration: docs/source/conf.py
 
-formats: pdf
+formats:
+  - pdf
 
 python:
   version: 3.6
