+++ tests/assets/generate_optim_assets.py
@@ -10,6 +10,7 @@ from pystiche.image.transforms import CaffePostprocessing, CaffePreprocessing
 from pystiche.ops import PixelComparisonOperator, TotalVariationOperator
 from pystiche.optim import default_image_optimizer, default_transformer_optimizer
 from pystiche.pyramid import ImagePyramid
+
 from utils import store_asset
 
 
