+++ tests/assets/generate_enc_assets.py
@@ -6,6 +6,7 @@ from torch import nn
 from torchvision import models
 
 import pystiche
+
 from utils import store_asset
 
 
