+++ tests/integration/core/test_home.py
@@ -5,20 +5,20 @@ from os import path
 import pystiche
 
 
-def test_home_default(self):
+def test_home_default():
     actual = pystiche.home()
     desired = path.expanduser(path.join("~", ".cache", "pystiche"))
-    self.assertEqual(actual, desired)
+    assert actual == desired
 
 
-def test_home_env(self):
+def test_home_env():
     tmp_dir = tempfile.mkdtemp()
     pystiche_home = os.getenv("PYSTICHE_HOME")
     os.environ["PYSTICHE_HOME"] = tmp_dir
     try:
         actual = pystiche.home()
         desired = tmp_dir
-        self.assertEqual(actual, desired)
+        assert actual == desired
     finally:
         if pystiche_home is None:
             del os.environ["PYSTICHE_HOME"]
