+++ tests/integration/core/test_objects.py
@@ -1,6 +1,9 @@
 import itertools
 from collections import OrderedDict
 
+import pytest
+import pytorch_testing_utils as ptu
+
 import torch
 
 import pystiche
@@ -9,15 +12,15 @@ from pystiche.misc import build_complex_obj_repr
 from tests.utils import skip_if_cuda_not_available
 
 
-def test_ComplexObject_repr_smoke(self):
+def test_ComplexObject_repr_smoke():
     class TestObject(pystiche.ComplexObject):
         pass
 
     test_object = TestObject()
-    self.assertIsInstance(repr(test_object), str)
+    assert isinstance(repr(test_object), str)
 
 
-def test_ComplexObject_repr(self):
+def test_ComplexObject_repr():
     _properties = OrderedDict((("a", 1),))
     extra_properties = OrderedDict((("b", 2),))
     _named_children = (("c", 3),)
@@ -51,10 +54,10 @@ def test_ComplexObject_repr(self):
     desired = build_complex_obj_repr(
         name="TestObject", properties=properties, named_children=named_children
     )
-    self.assertEqual(actual, desired)
+    assert actual == desired
 
 
-def test_LossDict_setitem_Tensor(self):
+def test_LossDict_setitem_Tensor():
     name = "loss"
     loss_dict = pystiche.LossDict()
 
@@ -63,18 +66,18 @@ def test_LossDict_setitem_Tensor(self):
 
     actual = loss_dict[name]
     desired = loss
-    self.assertTensorAlmostEqual(actual, desired)
+    ptu.assert_allclose(actual, desired)
 
 
-def test_LossDict_setitem_non_scalar_Tensor(self):
+def test_LossDict_setitem_non_scalar_Tensor():
     name = "loss"
     loss_dict = pystiche.LossDict()
 
-    with self.assertRaises(TypeError):
+    with pytest.raises(TypeError):
         loss_dict[name] = torch.ones(1)
 
 
-def test_LossDict_setitem_LossDict(self):
+def test_LossDict_setitem_LossDict():
     name = "loss"
     loss_dict = pystiche.LossDict()
     num_sub_losses = 3
@@ -87,18 +90,18 @@ def test_LossDict_setitem_LossDict(self):
     for idx in range(num_sub_losses):
         actual = loss_dict[f"{name}.{idx}"]
         desired = loss[str(idx)]
-        self.assertTensorAlmostEqual(actual, desired)
+        ptu.assert_allclose(actual, desired)
 
 
-def test_LossDict_setitem_other(self):
+def test_LossDict_setitem_other():
     name = "loss"
     loss_dict = pystiche.LossDict()
 
-    with self.assertRaises(TypeError):
+    with pytest.raises(TypeError):
         loss_dict[name] = 1.0
 
 
-def test_LossDict_aggregate_max_depth_gt_0(self):
+def test_LossDict_aggregate_max_depth_gt_0():
     def loss():
         return torch.tensor(1.0)
 
@@ -108,22 +111,22 @@ def test_LossDict_aggregate_max_depth_gt_0(self):
 
     actual = loss_dict.aggregate(1)
     desired = pystiche.LossDict((("0", 3 * loss()), ("1", loss())))
-    self.assertDictEqual(actual, desired)
+    ptu.assert_allclose(actual, desired)
 
     actual = loss_dict.aggregate(2)
     desired = pystiche.LossDict((("0.0", 2 * loss()), ("0.1", loss()), ("1", loss())))
-    self.assertDictEqual(actual, desired)
[…]
