+++ tox.ini
@@ -14,8 +14,10 @@ deps =
   pyimagetest
   pillow_affine
   dill
-  pytest-subtests
+  pytest-subtests >= 0.3.2
   coverage
+  # TODO: move to a released version
+  git+https://github.com/pmeier/pytorch_testing_utils
 commands =
   coverage run \
     --rcfile=.coveragerc \

[…]
