+++ tox.ini
@@ -9,21 +9,23 @@ envlist = tests
 isolated_build = True
 
 [testenv:tests]
+passenv =
+  GITHUB_ACTIONS
 deps =
   pytest >= 6
   pyimagetest
   pillow_affine
   dill
   pytest-subtests >= 0.3.2
-  coverage
+  pytest-cov
   # TODO: move to a released version
   git+https://github.com/pmeier/pytorch_testing_utils
 commands =
-  coverage run \
-    --rcfile=.coveragerc \
-    -m pytest \
-      -c pytest.ini \
-      {posargs}
+  pytest \
+    -c pytest.ini \
+    --cov=pystiche_papers \
+    --cov-report=xml \
+    --cov-config=.coveragerc
 
 [testenv:lint-style]
 whitelist_externals =
