### 2020-08-12 f5f3565
embrace pytest for core tests (#364)
* upgrade isort (#363)

* core

* fix
