### 2020-08-12 63d3421
embrace pytest for core tests (#364)
* upgrade isort (#363)

* core

* fix
