+++ tests/integration/core/test_objects.py
@@ -1,0 +1,297 @@
+import itertools
+from collections import OrderedDict
+
+import pytest
+import pytorch_testing_utils as ptu
+
+import torch
+
+import pystiche
+from pystiche.misc import build_complex_obj_repr
+
+from tests.utils import skip_if_cuda_not_available
+
+
+def test_ComplexObject_repr_smoke():
+    class TestObject(pystiche.ComplexObject):
+        pass
+
+    test_object = TestObject()
+    assert isinstance(repr(test_object), str)
+
+
+def test_ComplexObject_repr():
+    _properties = OrderedDict((("a", 1),))
+    extra_properties = OrderedDict((("b", 2),))
+    _named_children = (("c", 3),)
+    extra_named_children = (("d", 4),)
+
+    class TestObject(pystiche.ComplexObject):
+        def _properties(self):
+            return _properties
+
+        def extra_properties(self):
+            return extra_properties
+
+        def _named_children(self):
+            return iter(_named_children)
+
+        def extra_named_children(self):
+            return iter(extra_named_children)
+
+    test_object = TestObject()
+    properties = OrderedDict(
+        [
+            property
+            for property in itertools.chain(
+                _properties.items(), extra_properties.items()
+            )
+        ]
+    )
+    named_children = tuple(itertools.chain(_named_children, extra_named_children))
+
+    actual = repr(test_object)
+    desired = build_complex_obj_repr(
+        name="TestObject", properties=properties, named_children=named_children
+    )
+    assert actual == desired
+
+
+def test_LossDict_setitem_Tensor():
+    name = "loss"
+    loss_dict = pystiche.LossDict()
+
+    loss = torch.tensor(1.0)
+    loss_dict[name] = loss
+
+    actual = loss_dict[name]
+    desired = loss
+    ptu.assert_allclose(actual, desired)
+
+
+def test_LossDict_setitem_non_scalar_Tensor():
+    name = "loss"
+    loss_dict = pystiche.LossDict()
+
+    with pytest.raises(TypeError):
+        loss_dict[name] = torch.ones(1)
+
+
+def test_LossDict_setitem_LossDict():
+    name = "loss"
+    loss_dict = pystiche.LossDict()
+    num_sub_losses = 3
+
+    loss = pystiche.LossDict(
+        [(str(idx), torch.tensor(idx)) for idx in range(num_sub_losses)]
+    )
+    loss_dict[name] = loss
+
+    for idx in range(num_sub_losses):
+        actual = loss_dict[f"{name}.{idx}"]
+        desired = loss[str(idx)]
+        ptu.assert_allclose(actual, desired)
+
+
+def test_LossDict_setitem_other():
+    name = "loss"
+    loss_dict = pystiche.LossDict()
+
+    with pytest.raises(TypeError):
+        loss_dict[name] = 1.0
+
+
+def test_LossDict_aggregate_max_depth_gt_0():
+    def loss():
+        return torch.tensor(1.0)
+
+    loss_dict = pystiche.LossDict(
+        (("0.0.0", loss()), ("0.0.1", loss()), ("0.1", loss()), ("1", loss()))
+    )
+
+    actual = loss_dict.aggregate(1)
+    desired = pystiche.LossDict((("0", 3 * loss()), ("1", loss())))
+    ptu.assert_allclose(actual, desired)
+
[…]
