+++ tox.ini
@@ -23,9 +23,10 @@ deps =
 commands =
   pytest \
     -c pytest.ini \
-    --cov=pystiche_papers \
+    --cov=pystiche \
     --cov-report=xml \
-    --cov-config=.coveragerc
+    --cov-config=.coveragerc \
+    {posargs}
 
 [testenv:lint-style]
 whitelist_externals =
