### 2020-08-12 2cb25ed
use pytest-cov in favor of coverage (#365)
* fix coverage generation

* fix
