### 2020-08-12 db5f9f6
refactor tests for pystiche.data
