### 2020-08-12 cee8645
refactor tests for pystiche.data (#366)
