+++ tests/integration/data/test_collections.py
@@ -1,0 +1,258 @@
+import itertools
+import os
+import shutil
+from os import path
+
+import pytest
+import pytorch_testing_utils as ptu
+
+import torch
+from torch import nn
+
+from pystiche import data
+from pystiche.image import read_image, write_image
+
+
+def test_DownloadableImage_generate_file(subtests, test_image_url):
+    titles = (None, "girl with painted face")
+    authors = (None, "Steve Kelly")
+    titles_and_authors = itertools.product(titles, authors)
+
+    desireds = (
+        "test_image.png",
+        "steve_kelly.png",
+        "girl_with_painted_face.png",
+        "girl_with_painted_face__steve_kelly.png",
+    )
+
+    for (title, author), desired in zip(titles_and_authors, desireds):
+        with subtests.test(title=title, author=author):
+            actual = data.DownloadableImage.generate_file(
+                test_image_url, title=title, author=author
+            )
+            assert actual == desired
+
+
+def test_DownloadableImage_download(tmpdir, test_image_url, test_image):
+    image = data.DownloadableImage(test_image_url)
+    image.download(tmpdir)
+
+    file = path.join(tmpdir, image.file)
+    assert path.exists(file)
+
+    actual = read_image(file)
+    desired = test_image
+    ptu.assert_allclose(actual, desired)
+
+
+def test_DownloadableImage_download_guides(tmpdir, test_image_url, test_image):
+    guide = data.DownloadableImage(
+        test_image_url, file="guide" + path.splitext(test_image_url)[1],
+    )
+    image = data.DownloadableImage(
+        test_image_url, guides=data.DownloadableImageCollection({"guide": guide}),
+    )
+    image.download(tmpdir)
+
+    actual = read_image(path.join(tmpdir, guide.file))
+    desired = test_image
+    ptu.assert_allclose(actual, desired)
+
+
+def test_DownloadableImage_download_exist(
+    tmpdir, test_image_url, test_image_file, test_image
+):
+    image = data.DownloadableImage(test_image_url)
+    file = path.join(tmpdir, image.file)
+
+    shutil.copyfile(test_image_file, file)
+
+    with pytest.raises(FileExistsError):
+        image.download(tmpdir)
+
+    image.md5 = "invalid_hash"
+    with pytest.raises(FileExistsError):
+        image.download(tmpdir)
+
+    image.md5 = "a858d33c424eaac1322cf3cab6d3d568"
+    image.download(tmpdir)
+
+
+def test_DownloadableImage_download_overwrite(tmpdir, test_image_url, test_image):
+    def create_fake_image(file):
+        open(file, "wb").close()
+
+    image = data.DownloadableImage(test_image_url)
+    file = path.join(tmpdir, image.file)
+
+    create_fake_image(file)
+    image.download(tmpdir, overwrite=True)
+
+    actual = read_image(file)
+    desired = test_image
+    ptu.assert_allclose(actual, desired)
+
+    create_fake_image(file)
+    image.md5 = "a858d33c424eaac1322cf3cab6d3d568"
+    image.download(tmpdir, overwrite=True)
+
+    actual = read_image(file)
+    desired = test_image
+    ptu.assert_allclose(actual, desired)
+
+
+def test_DownloadableImage_read(tmpdir, test_image_url, test_image):
+    image = data.DownloadableImage(test_image_url)
+
+    actual = image.read(tmpdir)
+    desired = test_image
+    ptu.assert_allclose(actual, desired)
+
+
+def test_DownloadableImage_repr_smoke(test_image_url):
+    image = data.DownloadableImage(test_image_url)
+    assert isinstance(repr(image), str)
+
+
+def test_DownloadableImageCollection(tmpdir, test_image_url, test_image):
[…]
