### 2020-08-12 9c453e1
refactor tests for pystiche.data (#366)
