### 2020-08-12 ea1f172
refactor tests for pystiche.demo
