### 2020-08-12 601042d
refactor tests for pystiche.demo (#367)
