### 2020-08-12 a0d4f22
refactor tests for pystiche.enc
