+++ tests/test_enc.py
@@ -1,362 +1,0 @@
-from os import path
-
-import pytest
-
-import torch
-from torch import nn
-
-import pystiche
-from pystiche import enc
-from pystiche.image.transforms import CaffePreprocessing, TorchPreprocessing
-
-from .utils import ForwardPassCounter, PysticheTestCase
-
-
-class TestEncoder(PysticheTestCase):
-    def test_SequentialEncoder_call(self):
-        torch.manual_seed(0)
-        modules = (nn.Conv2d(3, 3, 3), nn.ReLU())
-        input = torch.rand(1, 3, 256, 256)
-
-        pystiche_encoder = enc.SequentialEncoder(modules)
-        torch_encoder = nn.Sequential(*modules)
-
-        actual = pystiche_encoder(input)
-        desired = torch_encoder(input)
-        self.assertTensorAlmostEqual(actual, desired)
-
-
-class TestModels(PysticheTestCase):
-    @pytest.mark.large_download
-    @pytest.mark.slow
-    @pytest.mark.flaky
-    def test_AlexNetMultiLayerEncoder(self):
-        asset = self.load_asset(path.join("enc", "alexnet"))
-
-        multi_layer_encoder = enc.alexnet_multi_layer_encoder(
-            weights="torch", preprocessing=False, allow_inplace=False
-        )
-        layers = tuple(multi_layer_encoder.children_names())
-        with torch.no_grad():
-            encs = multi_layer_encoder(asset.input.image, layers)
-
-        actual = dict(
-            zip(
-                layers,
-                [pystiche.TensorKey(x, precision=asset.params.precision) for x in encs],
-            )
-        )
-        desired = asset.output.enc_keys
-        self.assertDictEqual(actual, desired)
-
-    @pytest.mark.large_download
-    @pytest.mark.slow
-    def test_alexnet_multi_layer_encoder_smoke(self):
-        multi_layer_encoder = enc.alexnet_multi_layer_encoder()
-        self.assertIsInstance(multi_layer_encoder, enc.alexnet.AlexNetMultiLayerEncoder)
-
-    @pytest.mark.large_download
-    @pytest.mark.slow
-    @pytest.mark.flaky
-    def test_VGGMultiLayerEncoder(self):
-        archs = ("vgg11", "vgg13", "vgg16", "vgg19")
-        archs = (*archs, *[f"{arch}_bn" for arch in archs])
-
-        for arch in archs:
-            with self.subTest(arch=arch):
-                asset = self.load_asset(path.join("enc", arch))
-
-                get_vgg_multi_layer_encoder = enc.__dict__[
-                    f"{arch}_multi_layer_encoder"
-                ]
-                multi_layer_encoder = get_vgg_multi_layer_encoder(
-                    weights="torch", preprocessing=False, allow_inplace=False
-                )
-                layers = tuple(multi_layer_encoder.children_names())
-                with torch.no_grad():
-                    encs = multi_layer_encoder(asset.input.image, layers)
-
-                actual = dict(
-                    zip(
-                        layers,
-                        [
-                            pystiche.TensorKey(x, precision=asset.params.precision)
-                            for x in encs
-                        ],
-                    )
-                )
-                desired = asset.output.enc_keys
-                self.assertDictEqual(actual, desired)
-
-    @pytest.mark.large_download
-    @pytest.mark.slow
-    def test_vgg_multi_layer_encoder_smoke(self):
-        fns = (
-            enc.vgg11_multi_layer_encoder,
-            enc.vgg11_bn_multi_layer_encoder,
[…]
