+++ tests/integration/enc/test_multi_layer_encoder.py
@@ -1,0 +1,275 @@
+import pytest
+import pytorch_testing_utils as ptu
+
+import torch
+from torch import nn
+
+from pystiche import enc
+
+
+def test_MultiLayerEncoder():
+    modules = [(str(idx), nn.Module()) for idx in range(3)]
+    multi_layer_encoder = enc.MultiLayerEncoder(modules)
+
+    for name, module in modules:
+        actual = getattr(multi_layer_encoder, name)
+        desired = module
+        assert actual is desired
+
+
+def test_MultiLayerEncoder_named_children():
+    modules = [(str(idx), nn.Module()) for idx in range(3)]
+    multi_layer_encoder = enc.MultiLayerEncoder(modules)
+
+    actual = tuple(multi_layer_encoder.children_names())
+    desired = tuple(zip(*modules))[0]
+    assert actual == desired
+
+
+def test_MultiLayerEncoder_contains():
+    idcs = (0, 2)
+    modules = [(str(idx), nn.Module()) for idx in idcs]
+    multi_layer_encoder = enc.MultiLayerEncoder(modules)
+
+    for idx in idcs:
+        assert str(idx) in multi_layer_encoder
+
+    for idx in set(range(max(idcs))) - set(idcs):
+        assert str(idx) not in multi_layer_encoder
+
+
+def test_MultiLayerEncoder_extract_deepest_layer():
+    layers = [str(idx) for idx in range(3)]
+    modules = [(layer, nn.Module()) for layer in layers]
+    multi_layer_encoder = enc.MultiLayerEncoder(modules)
+
+    actual = multi_layer_encoder.extract_deepest_layer(layers)
+    desired = layers[-1]
+    assert actual == desired
+
+    actual = multi_layer_encoder.extract_deepest_layer(sorted(layers, reverse=True))
+    desired = layers[-1]
+    assert actual == desired
+
+    del multi_layer_encoder._modules[layers[-1]]
+
+    with pytest.raises(ValueError):
+        multi_layer_encoder.extract_deepest_layer(layers)
+
+    layers = layers[:-1]
+
+    actual = multi_layer_encoder.extract_deepest_layer(layers)
+    desired = layers[-1]
+    assert actual == desired
+
+
+def assert_named_modules_identical(actual, desired, equality_sufficient=False):
+    actual = tuple(actual)
+    desired = tuple(desired)
+
+    assert len(actual) == len(desired)
+    for (actual_name, actual_module), (desired_name, desired_module) in zip(
+        actual, desired
+    ):
+        assert actual_name == desired_name
+
+        if equality_sufficient:
+            assert actual_module == desired_module
+        else:
+            assert actual_module is desired_module
+
+
+def test_MultiLayerEncoder_named_children_to():
+    layers = [str(idx) for idx in range(3)]
+    modules = [(layer, nn.Module()) for layer in layers]
+    multi_layer_encoder = enc.MultiLayerEncoder(modules)
+
+    actuals = multi_layer_encoder.named_children_to(layers[-2])
+    desireds = modules[:-2]
+    assert_named_modules_identical(actuals, desireds)
+
+    actuals = multi_layer_encoder.named_children_to(layers[-2], include_last=True)
+    desireds = modules[:-1]
+    assert_named_modules_identical(actuals, desireds)
+
+
+def test_MultiLayerEncoder_named_children_from():
+    layers = [str(idx) for idx in range(3)]
+    modules = [(layer, nn.Module()) for layer in layers]
+    multi_layer_encoder = enc.MultiLayerEncoder(modules)
+
[…]
