### 2020-08-12 c8ecbc7
refactor tests for pystiche.enc (#368)
