### 2020-08-12 fecae54
refactor tests for pystiche.enc (#368)
