+++ tests/assets/__init__.py
@@ -3,11 +3,13 @@ import os
 from os import path
 from urllib.parse import urljoin
 
+import dill
+
 from pystiche.image import read_image as _read_image
 
 HERE = path.abspath(path.dirname(__file__))
 
-__all__ = ["get_image_url", "get_image_file", "read_image"]
+__all__ = ["get_image_url", "get_image_file", "read_image", "load_asset"]
 
 
 def get_image_file(name):
@@ -29,3 +31,11 @@ _read_image = functools.lru_cache()(_read_image)
 def read_image(name):
     image = _read_image(get_image_file(name))
     return image.clone()
+
+
+@functools.lru_cache()
+def load_asset(type, name, ext=".asset"):
+    file = path.join(HERE, type, path.splitext(name)[0] + ext)
+
+    with open(file, "rb") as fh:
+        return dill.load(fh)
