### 2020-08-12 a927a3e
refactor tests for pystiche.loss
