### 2020-08-12 57b8de0
refactor tests for pystiche.loss (#369)
* refactor tests for pystiche.loss

* cleanup
