### 2020-08-12 7c3b516
refactor tests for pystiche.meta (#370)
