### 2020-08-12 d80d6cb
refactor tests for pystiche.misc (#371)
