+++ tests/test_misc.py
@@ -1,205 +1,0 @@
-import math
-from os import path
-
-import torch
-
-from pystiche import misc
-
-from .utils import PysticheTestCase, get_tmp_dir
-
-
-class TestMisc(PysticheTestCase):
-    def test_prod(self):
-        n = 10
-        iterable = range(1, n + 1)
-
-        actual = misc.prod(iterable)
-        desired = math.factorial(n)
-        self.assertEqual(actual, desired)
-
-    def test_to_1d_arg(self):
-        val = 0
-        actual = misc.to_1d_arg(val)
-        desired = (val,)
-        self.assertTupleEqual(actual, desired)
-
-        val = (0,)
-        actual = misc.to_1d_arg(val)
-        desired = val
-        self.assertTupleEqual(actual, desired)
-
-        val = 0
-        actual = misc.to_1d_arg([val])
-        desired = (val,)
-        self.assertTupleEqual(actual, desired)
-
-        val = (0, 0)
-        with self.assertRaises(RuntimeError):
-            misc.to_1d_arg(val)
-
-    def test_to_2d_arg(self):
-        val = 0
-        actual = misc.to_2d_arg(val)
-        desired = (val, val)
-        self.assertTupleEqual(actual, desired)
-
-        val = (0, 0)
-        actual = misc.to_2d_arg(val)
-        desired = val
-        self.assertTupleEqual(actual, desired)
-
-        val = 0
-        actual = misc.to_2d_arg([val] * 2)
-        desired = (val, val)
-        self.assertTupleEqual(actual, desired)
-
-        val = (0,)
-        with self.assertRaises(RuntimeError):
-            misc.to_2d_arg(val)
-
-    def test_to_3d_arg(self):
-        val = 0
-        actual = misc.to_3d_arg(val)
-        desired = (val, val, val)
-        self.assertTupleEqual(actual, desired)
-
-        val = (0, 0, 0)
-        actual = misc.to_3d_arg(val)
-        desired = val
-        self.assertTupleEqual(actual, desired)
-
-        val = 0
-        actual = misc.to_3d_arg([val] * 3)
-        desired = (val, val, val)
-        self.assertTupleEqual(actual, desired)
-
-        val = (0,)
-        with self.assertRaises(RuntimeError):
-            misc.to_3d_arg(val)
-
-    def test_zip_equal(self):
-        foo = (1, 2)
-        bar = ("a", "b")
-
-        actual = tuple(misc.zip_equal(foo, bar))
-        desired = tuple(zip(foo, bar))
-        self.assertTupleEqual(actual, desired)
-
-        foo = (1, 2)
-        bar = ("a", "b", "c")
-
-        with self.assertRaises(RuntimeError):
-            misc.zip_equal(foo, bar)
-
-    def test_verify_str_arg(self):
-        arg = None
-        with self.assertRaises(ValueError):
-            misc.verify_str_arg(arg)
-
-        arg = "foo"
-        valid_args = ("bar", "baz")
-        with self.assertRaises(ValueError):
-            misc.verify_str_arg(arg, valid_args=valid_args)
-
-        arg = "foo"
-        valid_args = ("foo", "bar")
-
-        actual = misc.verify_str_arg(arg, valid_args=valid_args)
-        desired = arg
-        self.assertEqual(actual, desired)
-
-    def assertScalarTensorAlmostEqual(self, actual, desired):
-        self.assertAlmostEqual(actual.item(), desired.item())
-
-    def test_get_input_image_tensor(self):
-        image = torch.tensor(0.0)
-
-        starting_point = image
-        actual = misc.get_input_image(starting_point)
-        desired = image
-        self.assertIsNot(actual, desired)
[…]
