### 2020-08-12 b175182
refactor tests for pystiche.ops (#372)
