### 2020-08-12 63890de
refactor tests for pystiche.ops (#372)
