+++ tests/test_ops.py
@@ -1,1019 +1,0 @@
-import itertools
-
-import torch
-from torch import nn
-from torch.nn.functional import mse_loss
-
-import pystiche
-from pystiche import ops
-from pystiche.enc import MultiLayerEncoder, SequentialEncoder, SingleLayerEncoder
-from pystiche.ops import functional as F
-
-from .utils import PysticheTestCase
-
-
-class TestComparison(PysticheTestCase):
-    def test_FeatureReconstructionOperator_call(self):
-        torch.manual_seed(0)
-        target_image = torch.rand(1, 3, 128, 128)
-        input_image = torch.rand(1, 3, 128, 128)
-        encoder = SequentialEncoder((nn.Conv2d(3, 3, 1),))
-
-        op = ops.FeatureReconstructionOperator(encoder)
-        op.set_target_image(target_image)
-
-        actual = op(input_image)
-        desired = mse_loss(encoder(input_image), encoder(target_image))
-        self.assertTensorAlmostEqual(actual, desired)
-
-    def test_GramOperator_call(self):
-        torch.manual_seed(0)
-        target_image = torch.rand(1, 3, 128, 128)
-        input_image = torch.rand(1, 3, 128, 128)
-        encoder = SequentialEncoder((nn.Conv2d(3, 3, 1),))
-
-        op = ops.GramOperator(encoder)
-        op.set_target_image(target_image)
-
-        actual = op(input_image)
-        desired = mse_loss(
-            pystiche.gram_matrix(encoder(input_image), normalize=True),
-            pystiche.gram_matrix(encoder(target_image), normalize=True),
-        )
-        self.assertTensorAlmostEqual(actual, desired)
-
-    def test_MRFOperator_scale_and_rotate_transforms(self):
-        num_scale_steps = 1
-        scale_step_width = 10e-2
-        num_rotate_steps = 1
-        rotate_step_width = 30.0
-
-        target_transforms = ops.MRFOperator.scale_and_rotate_transforms(
-            num_scale_steps=num_scale_steps,
-            scale_step_width=scale_step_width,
-            num_rotate_steps=num_rotate_steps,
-            rotate_step_width=rotate_step_width,
-        )
-        self.assertEqual(
-            len(target_transforms),
-            (num_scale_steps * 2 + 1) * (num_rotate_steps * 2 + 1),
-        )
-
-        actual = {
-            (transform.scaling_factor, transform.rotation_angle)
-            for transform in target_transforms
-        }
-        desired = set(
-            itertools.product(
-                (1.0 - scale_step_width, 1.0, 1.0 + scale_step_width),
-                (-rotate_step_width, 0.0, rotate_step_width),
-            )
-        )
-        self.assertSetEqual(actual, desired)
-
-    def test_MRFOperator_enc_to_repr_guided(self):
-        class Identity(pystiche.Module):
-            def forward(self, image):
-                return image
-
-        patch_size = 2
-        stride = 2
-
-        op = ops.MRFOperator(
-            SequentialEncoder((Identity(),)), patch_size, stride=stride
-        )
-
-        with self.subTest(enc="constant"):
-            enc = torch.ones(1, 4, 8, 8)
-
-            actual = op.enc_to_repr(enc, is_guided=True)
-            desired = torch.ones(0, 4, stride, stride)
-            self.assertTensorAlmostEqual(actual, desired)
-
-        with self.subTest(enc="spatial_mix"):
-            constant = torch.ones(1, 4, 4, 8)
-            varying = torch.rand(1, 4, 4, 8)
-            enc = torch.cat((constant, varying), dim=2)
-
-            actual = op.enc_to_repr(enc, is_guided=True)
[…]
