+++ tests/integration/ops/test_op.py
@@ -1,0 +1,485 @@
+import pytest
+import pytorch_testing_utils as ptu
+
+import torch
+from torch import nn
+
+from pystiche import enc, ops
+
+
+def test_Operator_set_input_guide():
+    class TestOperator(ops.Operator):
+        def process_input_image(self, image):
+            pass
+
+    torch.manual_seed(0)
+    guide = torch.rand(1, 1, 32, 32)
+
+    test_op = TestOperator()
+    assert not test_op.has_input_guide
+
+    test_op.set_input_guide(guide)
+    assert test_op.has_input_guide
+
+    actual = test_op.input_guide
+    desired = guide
+    ptu.assert_allclose(actual, desired)
+
+
+def test_Operator_apply_guide():
+    torch.manual_seed(0)
+    image = torch.rand(1, 3, 32, 32)
+    guide = torch.rand(1, 1, 32, 32)
+
+    actual = ops.Operator.apply_guide(image, guide)
+    desired = image * guide
+    ptu.assert_allclose(actual, desired)
+
+
+def test_Operator_named_operators():
+    # TODO: add
+    pass
+
+
+def test_Operator_call():
+    class TestOperator(ops.Operator):
+        def process_input_image(self, image):
+            return image + 1.0
+
+    torch.manual_seed(0)
+    image = torch.rand(1, 3, 128, 128)
+
+    test_op = TestOperator(score_weight=2.0)
+
+    actual = test_op(image)
+    desired = (image + 1.0) * 2.0
+    ptu.assert_allclose(actual, desired)
+
+
+def test_PixelRegularizationOperator_call():
+    class TestOperator(ops.PixelRegularizationOperator):
+        def input_image_to_repr(self, image):
+            return image * 2.0
+
+        def calculate_score(self, input_repr):
+            return input_repr + 1.0
+
+    torch.manual_seed(0)
+    image = torch.rand(1, 3, 128, 128)
+
+    test_op = TestOperator()
+
+    actual = test_op(image)
+    desired = image * 2.0 + 1.0
+    ptu.assert_allclose(actual, desired)
+
+
+def test_PixelRegularizationOperator_call_guided():
+    class TestOperator(ops.PixelRegularizationOperator):
+        def input_image_to_repr(self, image):
+            return image * 2.0
+
+        def calculate_score(self, input_repr):
+            return input_repr + 1.0
+
+    torch.manual_seed(0)
+    image = torch.rand(1, 3, 32, 32)
+    guide = torch.rand(1, 1, 32, 32)
+
+    test_op = TestOperator()
+    test_op.set_input_guide(guide)
+
+    actual = test_op(image)
+    desired = TestOperator.apply_guide(image, guide) * 2.0 + 1.0
+    ptu.assert_allclose(actual, desired)
+
+
+def test_EncodingRegularizationOperator_call():
+    class TestOperator(ops.EncodingRegularizationOperator):
+        def input_enc_to_repr(self, image):
+            return image * 2.0
+
+        def calculate_score(self, input_repr):
+            return input_repr + 1.0
+
+    torch.manual_seed(0)
+    image = torch.rand(1, 3, 128, 128)
+    encoder = enc.SequentialEncoder((nn.Conv2d(3, 3, 1),))
+
+    test_op = TestOperator(encoder)
+
+    actual = test_op(image)
+    desired = encoder(image) * 2.0 + 1.0
+    ptu.assert_allclose(actual, desired)
+
+
+def test_EncodingRegularizationOperator_call_guided():
+    class TestOperator(ops.EncodingRegularizationOperator):
+        def input_enc_to_repr(self, image):
+            return image * 2.0
+
+        def calculate_score(self, input_repr):
+            return input_repr + 1.0
+
[…]

[…]
