### 2020-08-13 80655ad
refactor tests for pystiche.optim
