### 2020-08-13 86df804
refactor tests for pystiche.pyramid
