### 2020-08-13 48019e9
refactor tests for pystiche.optim (#373)
* refactor tests for pystiche.optim

* delete old
