### 2020-08-13 2681e68
refactor tests for pystiche.optim (#373)
* refactor tests for pystiche.optim

* delete old
