+++ tests/integration/optim/test_meter.py
@@ -1,0 +1,258 @@
+from datetime import datetime, timedelta
+
+import pytest
+
+import torch
+
+import pystiche
+from pystiche import optim
+
+
+class FloatMeter(optim.FloatMeter):
+    def __init__(self, window_size=10):
+        super().__init__("float_meter", window_size=window_size)
+
+    def __str__(self):
+        pass
+
+
+def test_FloatMeter_count():
+    torch.manual_seed(0)
+    vals = torch.rand(10)
+
+    meter = FloatMeter()
+    meter.update(vals.tolist())
+
+    actual = meter.count
+    desired = len(vals)
+    assert actual == desired
+
+
+def test_FloatMeter_last_val():
+    torch.manual_seed(0)
+    vals = torch.rand(10)
+
+    meter = FloatMeter()
+    meter.update(vals.tolist())
+
+    actual = meter.last_val
+    desired = vals[-1].item()
+    assert actual == pytest.approx(desired)
+
+
+def test_FloatMeter_global_sum():
+    torch.manual_seed(0)
+    vals = torch.rand(10)
+
+    meter = FloatMeter()
+    meter.update(vals.tolist())
+
+    actual = meter.global_sum
+    desired = torch.sum(vals).item()
+    assert actual == pytest.approx(desired)
+
+
+def test_FloatMeter_global_min():
+    torch.manual_seed(0)
+    vals = torch.rand(10)
+
+    meter = FloatMeter()
+    meter.update(vals.tolist())
+
+    actual = meter.global_min
+    desired = torch.min(vals).item()
+    assert actual == pytest.approx(desired)
+
+
+def test_FloatMeter_global_max():
+    torch.manual_seed(0)
+    vals = torch.rand(10)
+
+    meter = FloatMeter()
+    meter.update(vals.tolist())
+
+    actual = meter.global_max
+    desired = torch.max(vals).item()
+    assert actual == pytest.approx(desired)
+
+
+def test_FloatMeter_global_avg():
+    torch.manual_seed(0)
+    vals = torch.rand(10)
+
+    meter = FloatMeter()
+    meter.update(vals.tolist())
+
+    actual = meter.global_avg
+    desired = torch.mean(vals).item()
+    assert actual == pytest.approx(desired)
+
+
+def test_FloatMeter_global_avg_empty():
+    meter = FloatMeter()
+    with pytest.raises(RuntimeError):
+        meter.global_avg
+
+
+def test_FloatMeter_local_avg():
+    window_size = 5
+
+    torch.manual_seed(0)
+    vals = torch.rand(10)
+
+    meter = FloatMeter(window_size=window_size)
+    meter.update(vals.tolist())
+
+    actual = meter.local_avg
+    desired = torch.mean(vals[-window_size:]).item()
+    assert actual == pytest.approx(desired)
+
+
+def test_FloatMeter_local_avg_empty():
+    meter = FloatMeter()
+    with pytest.raises(RuntimeError):
+        meter.local_avg
+
+
+def test_AverageMeter_update_tensor():
+    actual_meter = optim.AverageMeter("actual_meter")
+    desired_meter = optim.AverageMeter("desired_meter")
+
+    torch.manual_seed(0)
+    vals = torch.rand(10)
+
+    actual_meter.update(vals)
+    desired_meter.update(vals.tolist())
+
+    for attr in (
+        "count",
+        "last_val",
+        "global_sum",
+        "global_min",
+        "global_max",
+        "global_avg",
+        "local_avg",
+    ):
+        actual = getattr(actual_meter, attr)
+        desired = getattr(desired_meter, attr)
+        assert actual == pytest.approx(desired)
+
+
+def test_AverageMeter_str_smoke():
+    for show_local_avg in (True, False):
[…]
