### 2020-08-13 e6e1c61
refactor tests for pystiche.pyramid (#374)
