### 2020-08-14 febcbf0
limit pyimagetest to < 0.3
