+++ tox.ini
@@ -13,7 +13,7 @@ passenv =
   GITHUB_ACTIONS
 deps =
   pytest >= 6
-  pyimagetest
+  pyimagetest < 0.3
   pillow_affine
   dill
   pytest-subtests >= 0.3.2
