### 2020-08-14 ee69888
refactor tests for pystiche.image
