### 2020-08-14 8be881b
refactor tests for pystiche.image.transforms
