+++ tests/test_image_transforms.py
@@ -1,779 +1,0 @@
-import numpy as np
-import pillow_affine as pa
-from PIL import Image
-
-import torch
-
-from pystiche.image import (
-    calculate_aspect_ratio,
-    edge_to_image_size,
-    extract_edge_size,
-    extract_image_size,
-    io,
-    is_single_image,
-    make_batched_image,
-    make_single_image,
-    transforms,
-)
-
-from .utils import PysticheTestCase
-
-
-class PysticheTransfromTestCase(PysticheTestCase):
-    def assertIdentityTransform(self, transform, image=None, tolerance=1e-2):
-        if image is None:
-            image = self.load_image()
-        actual = image
-        desired = transform(image)
-        self.assertImagesAlmostEqual(actual, desired, tolerance=tolerance)
-
-    @staticmethod
-    def get_pil_affine_transform(affine_transform, expand=False):
-        def transform(image):
-            transform_params = affine_transform.extract_transform_params(
-                image.size, expand=expand
-            )
-            return image.transform(*transform_params, resample=Image.BILINEAR)
-
-        return transform
-
-    def assertTransformEqualsPIL(
-        self,
-        pystiche_transform,
-        pil_transform,
-        pystiche_image=None,
-        pil_image=None,
-        tolerance=1e-2,
-    ):
-        def parse_images(pystiche_image, pil_image):
-            if pystiche_image is None and pil_image is None:
-                pystiche_image = self.load_image(backend="pystiche")
-                pil_image = self.load_image(backend="PIL")
-            elif pystiche_image is None:
-                pystiche_image = io.import_from_pil(pil_image)
-            elif pil_image is None:
-                pil_image = io.export_to_pil(pystiche_image)
-
-            return pystiche_image, pil_image
-
-        def get_single_and_batched_pystiche_images(pystiche_image):
-            if is_single_image(pystiche_image):
-                pystiche_single_image = pystiche_image
-                pystiche_batched_image = make_batched_image(pystiche_single_image)
-            else:
-                pystiche_batched_image = pystiche_image
-                pystiche_single_image = make_single_image(pystiche_batched_image)
-
-            return pystiche_single_image, pystiche_batched_image
-
-        def assert_transform_equality(pystiche_image, pil_image):
-            actual = pystiche_transform(pystiche_image)
-            desired = pil_transform(pil_image)
-            self.assertImagesAlmostEqual(actual, desired, tolerance=tolerance)
-
-        pystiche_image, pil_image = parse_images(pystiche_image, pil_image)
-
-        (
-            pystiche_single_image,
-            pystiche_batched_image,
-        ) = get_single_and_batched_pystiche_images(pystiche_image)
-
-        assert_transform_equality(pystiche_single_image, pil_image)
-        assert_transform_equality(pystiche_batched_image, pil_image)
-
-
-class TestColor(PysticheTransfromTestCase):
-    def test_RGBToGrayscale(self):
-        def PILRGBToGrayscale():
-            def transform(image):
-                assert image.mode == "RGB"
-                return image.convert("L")
-
-            return transform
-
-        self.assertTransformEqualsPIL(
-            pystiche_transform=transforms.RGBToGrayscale(),
-            pil_transform=PILRGBToGrayscale(),
-        )
-
-    def test_GrayscaleToFakegrayscale(self):
-        def PILGrayscaleToFakegrayscale():
-            def transform(image):
-                assert image.mode == "L"
-                return image.convert("RGB")
-
-            return transform
-
-        self.assertTransformEqualsPIL(
[…]
