+++ .github/workflows/tests.yml
@@ -20,7 +20,7 @@ on:
       - ".github/workflows/tests.yml"
 
 jobs:
-  unit:
+  integration:
     strategy:
       matrix:
         os: [ubuntu-latest, windows-latest, macos-latest]
