+++ tests/utils.py
@@ -3,139 +3,15 @@ import os
 import shutil
 import stat
 import tempfile
-from os import path
 
-import dill
-import numpy as np
-import pyimagetest
 import pytest
-from PIL import Image
 
 import torch
-from torch import nn
 
 __all__ = [
-    "PysticheTestCase",
     "get_tmp_dir",
     "skip_if_cuda_not_available",
-    "ForwardPassCounter",
 ]
-
-
-class PysticheImageBackend(pyimagetest.ImageBackend):
-    @property
-    def native_image_type(self):
-        return torch.Tensor
-
-    def import_image(self, file):
-        pil_image = Image.open(file)
-        np_image = np.array(pil_image, dtype=np.float32) / 255.0
-
-        pystiche_image = torch.from_numpy(np_image)
-        if pystiche_image.dim() == 2:
-            pystiche_image = pystiche_image.unsqueeze(0)
-        elif pystiche_image.dim() == 3:
-            pystiche_image = pystiche_image.permute((2, 0, 1))
-        pystiche_image = pystiche_image.unsqueeze(0)
-
-        return pystiche_image
-
-    def export_image(self, image):
-        image = image.detach().cpu()
-        if image.dim() == 4 and image.size()[0] == 1:
-            image = image.squeeze(0)
-        return image.permute((1, 2, 0)).numpy()
-
-
-class PysticheTestCase(pyimagetest.ImageTestCase):
-    project_root = path.abspath(path.join(path.dirname(__file__), ".."))
-    package_name = "pystiche"
-
-    @property
-    def package_root(self):
-        return path.join(self.project_root, self.package_name)
-
-    @property
-    def test_root(self):
-        return path.join(self.project_root, "tests")
-
-    @property
-    def test_assets_root(self):
-        return path.join(self.test_root, "assets")
-
-    def setUp(self):
-        # the builtin torchvision backend is removed, since it operates on the same
-        # native image type (torch.Tensor), which renders the automatic inference of
-        # the image backend impossible
-        self.remove_image_backend("torchvision")
-        self.add_image_backend(self.package_name, PysticheImageBackend())
-
-    def default_image_backend(self) -> str:
-        return self.package_name
-
-    def default_image_file(self) -> str:
-        return path.join(self.test_assets_root, "image", "test_image.png")
-
-    def load_batched_image(self, batch_size=1, file=None):
-        return self.load_image(file=file).repeat(batch_size, 1, 1, 1)
-
-    def load_single_image(self, file=None):
-        return self.load_batched_image(file=file).squeeze(0)
-
-    def assertFloatAlmostEqual(self, actual, desired, **kwargs):
-        if isinstance(actual, torch.Tensor):
-            actual = actual.item()
-
-        if isinstance(desired, torch.Tensor):
-            desired = desired.item()
-
-        np.testing.assert_almost_equal(actual, desired, **kwargs)
-
-    def assertTensorAlmostEqual(
-        self, actual, desired, check_dtype=True, check_device=True, **kwargs
-    ):
-        def cast(x):
-            return x.detach().cpu().numpy()
-
-        if check_dtype:
-            self.assertEqual(actual.dtype, desired.dtype)
-
-        if check_device:
-            self.assertEqual(actual.device, desired.device)
-
-        np.testing.assert_allclose(cast(actual), cast(desired), **kwargs)
-
-    def assertNamedChildrenEqual(self, actuals, desireds, equality_sufficient=False):
-        self.assertCountEqual(actuals, desireds)
-        for actual, desired in zip(actuals, desireds):
-            actual_name, actual_module = actual
-            desired_name, desired_module = actual
[…]
