+++ .github/workflows/publish.yml
@@ -32,7 +32,7 @@ jobs:
         run: python -m pep517.build --source --binary .
 
 #      - name: Upload to PyPI
-#        enc:
+#        env:
 #          TWINE_REPOSITORY: pypi
 #          TWINE_USERNAME: ${{ secrets.PYPI_USERNAME }}
 #          TWINE_PASSWORD: ${{ secrets.PYPI_PASSWORD }}
