### 2020-08-14 8d4a082
Embrace pytest paradigm for test suite (#380)
* embrace pytest for core tests (#364)

* upgrade isort (#363)

* core

* fix

* use pytest-cov in favor of coverage (#365)

* fix coverage generation

* fix

* refactor tests for pystiche.data (#366)

* refactor tests for pystiche.demo (#367)

* refactor tests for pystiche.enc (#368)

* refactor tests for pystiche.loss (#369)

* refactor tests for pystiche.loss

* cleanup

* refactor tests for pystiche.meta (#370)

* refactor tests for pystiche.misc (#371)

* refactor tests for pystiche.ops (#372)

* refactor tests for pystiche.optim (#373)

* ref
[…]
