+++ tests/integration/ops/test_comparison.py
@@ -1,0 +1,208 @@
+import itertools
+
+import pytorch_testing_utils as ptu
+
+import torch
+from torch import nn
+from torch.nn.functional import mse_loss
+
+import pystiche
+import pystiche.ops.functional as F
+from pystiche import enc, ops
+
+
+def test_FeatureReconstructionOperator_call():
+    torch.manual_seed(0)
+    target_image = torch.rand(1, 3, 128, 128)
+    input_image = torch.rand(1, 3, 128, 128)
+    encoder = enc.SequentialEncoder((nn.Conv2d(3, 3, 1),))
+
+    op = ops.FeatureReconstructionOperator(encoder)
+    op.set_target_image(target_image)
+
+    actual = op(input_image)
+    desired = mse_loss(encoder(input_image), encoder(target_image))
+    ptu.assert_allclose(actual, desired)
+
+
+def test_GramOperator_call():
+    torch.manual_seed(0)
+    target_image = torch.rand(1, 3, 128, 128)
+    input_image = torch.rand(1, 3, 128, 128)
+    encoder = enc.SequentialEncoder((nn.Conv2d(3, 3, 1),))
+
+    op = ops.GramOperator(encoder)
+    op.set_target_image(target_image)
+
+    actual = op(input_image)
+    desired = mse_loss(
+        pystiche.gram_matrix(encoder(input_image), normalize=True),
+        pystiche.gram_matrix(encoder(target_image), normalize=True),
+    )
+    ptu.assert_allclose(actual, desired)
+
+
+def test_MRFOperator_scale_and_rotate_transforms():
+    num_scale_steps = 1
+    scale_step_width = 10e-2
+    num_rotate_steps = 1
+    rotate_step_width = 30.0
+
+    target_transforms = ops.MRFOperator.scale_and_rotate_transforms(
+        num_scale_steps=num_scale_steps,
+        scale_step_width=scale_step_width,
+        num_rotate_steps=num_rotate_steps,
+        rotate_step_width=rotate_step_width,
+    )
+    assert len(target_transforms) == (num_scale_steps * 2 + 1) * (
+        num_rotate_steps * 2 + 1
+    )
+
+    actual = {
+        (transform.scaling_factor, transform.rotation_angle)
+        for transform in target_transforms
+    }
+    desired = set(
+        itertools.product(
+            (1.0 - scale_step_width, 1.0, 1.0 + scale_step_width),
+            (-rotate_step_width, 0.0, rotate_step_width),
+        )
+    )
+    assert actual == desired
+
+
+def test_MRFOperator_enc_to_repr_guided(subtests):
+    class Identity(pystiche.Module):
+        def forward(self, image):
+            return image
+
+    patch_size = 2
+    stride = 2
+
+    op = ops.MRFOperator(
+        enc.SequentialEncoder((Identity(),)), patch_size, stride=stride
+    )
+
+    with subtests.test(enc="constant"):
+        enc_ = torch.ones(1, 4, 8, 8)
+
+        actual = op.enc_to_repr(enc_, is_guided=True)
+        desired = torch.ones(0, 4, stride, stride)
+        ptu.assert_allclose(actual, desired)
+
+    with subtests.test(enc="spatial_mix"):
+        constant = torch.ones(1, 4, 4, 8)
+        varying = torch.rand(1, 4, 4, 8)
+        enc_ = torch.cat((constant, varying), dim=2)
+
+        actual = op.enc_to_repr(enc_, is_guided=True)
+        desired = pystiche.extract_patches2d(varying, patch_size, stride=stride)
[…]
