### 2020-08-14 9a24494
fix image.edge_to_image_size for horz edge
