+++ setup.cfg
@@ -38,6 +38,7 @@ python_requires = >=3.6
 install_requires =
     torch==1.6
     torchvision==0.7
+    kornia==0.4.1
     pillow
     numpy
     tqdm
