### 2020-08-14 pmeier/pystiche#382 @pmeier
Currently the following things are missing from `kornia` that are needed to replace `pystiche.image.transforms`

- [ ] An affine transformation. `kornia.geometry.affine` is a available as functional, but we need it as transformation.

For each item on this list we need to either send PRs to `kornia` and fix this upstream or provide compatibility functions.
