### 2020-08-17 f539308
split tests in integration and examples
