+++ tox.ini
@@ -14,7 +14,8 @@ deps =
   pytest-subtests >= 0.3.2
   pytest-cov
   pytest-mock >= 3.1
-    dill
+  dill
+  gitignore_parser
 commands = pytest -c pytest.ini --durations=25
 
 [testenv:tests-integration]
@@ -26,7 +27,6 @@ deps =
   pillow_affine
   pyimagetest >= 0.3
   pillow_affine
-  dill
   # TODO: move to a released version
   git+https://github.com/pmeier/pytorch_testing_utils
 commands =
@@ -44,7 +44,6 @@ deps =
   # Additional sphinx-gallery dependencies
   # https://sphinx-gallery.github.io/stable/index.html#install-via-pip
   matplotlib
-  gitignore_parser
 commands =
   {[tests-common]commands} \
     {posargs:tests/galleries}
