+++ tox.ini
@@ -14,7 +14,6 @@ deps =
   pytest-subtests >= 0.3.2
   pytest-cov
   pytest-mock >= 3.1
-  dill
   gitignore_parser
 commands = pytest -c pytest.ini --durations=25
 
@@ -27,6 +26,7 @@ deps =
   pillow_affine
   pyimagetest >= 0.3
   pillow_affine
+  dill
   # TODO: move to a released version
   git+https://github.com/pmeier/pytorch_testing_utils
 commands =
@@ -34,19 +34,18 @@ commands =
     --cov=pystiche \
     --cov-report=xml \
     --cov-config=.coveragerc \
-    {posargs:tests/integration}
+    tests/integration \
+    {posargs}
 
 [testenv:tests-galleries]
 deps =
   {[tests-common]deps}
   importlib_metadata
-  sphinx-gallery>=0.7.0
-  # Additional sphinx-gallery dependencies
-  # https://sphinx-gallery.github.io/stable/index.html#install-via-pip
-  matplotlib
+  {[sphinx-gallery]deps}
 commands =
   {[tests-common]commands} \
-    {posargs:tests/galleries}
+    tests/galleries \
+    {posargs}
 
 [testenv:lint-style]
 whitelist_externals =
@@ -63,6 +62,14 @@ deps =
 commands =
   mypy --config-file=mypy.ini
 
+[sphinx-gallery]
+deps =
+  sphinx-gallery>=0.7.0
+  # Additional sphinx-gallery dependencies
+  # https://sphinx-gallery.github.io/stable/index.html#install-via-pip
+  matplotlib
+
+
 [docs-common]
 passenv =
   HOME
@@ -77,10 +84,7 @@ deps =
   importlib-metadata
   sphinxcontrib-bibtex
   sphinx_autodoc_typehints >= 1.11
-  sphinx-gallery>=0.7.0
-  # Additional sphinx-gallery dependencies
-  # https://sphinx-gallery.github.io/stable/index.html#install-via-pip
-  matplotlib
+  {[sphinx-gallery]deps}
   sphinx_rtd_theme
 changedir = docs
 
