+++ tox.ini
@@ -14,6 +14,7 @@ deps =
   pytest-subtests >= 0.3.2
   pytest-cov
   pytest-mock >= 3.1
+  dill
   gitignore_parser
 commands = pytest -c pytest.ini --durations=25
 
@@ -26,7 +27,6 @@ deps =
   pillow_affine
   pyimagetest >= 0.3
   pillow_affine
-  dill
   # TODO: move to a released version
   git+https://github.com/pmeier/pytorch_testing_utils
 commands =
