### 2020-08-18 1e1f01b
fix parse_gitignore for pystiche.home()
