+++ tox.ini
@@ -15,7 +15,7 @@ deps =
   pytest-cov
   pytest-mock >= 3.1
   dill
-  gitignore_parser
+  igittigitt
 commands = pytest -c pytest.ini --durations=25
 
 [testenv:tests-integration]
