+++ pyproject.toml
@@ -31,6 +31,7 @@ filter_files = true
 extra_standard_library = ["importlib_metadata"]
 known_third_party = [
   "dill",
+  "igittigitt",
   "light-the-torch",
   "matplotlib",
   "numpy",
