### 2020-08-18 5ce7ee4
Add smoke tests for galleries (#383)
* split tests in integration and examples

* add workflow

* fix dependencies

* fix tox invocation in workflows

* fix

* cleanup

* fix parse_gitignore for pystiche.home()

* replace gitignore_parser with igittigitt
