+++ tests/fixtures.py
@@ -4,7 +4,21 @@ import pytest
 
 from torch import nn
 
-from . import assets
+import pystiche
+
+from . import assets, utils
+
+
+@pytest.fixture(scope="session", autouse=True)
+def watch_project_dir():
+    with utils.watch_dir("."):
+        yield
+
+
+@pytest.fixture(scope="session", autouse=True)
+def watch_pystiche_home():
+    with utils.watch_dir(pystiche.home()):
+        yield
 
 
 @pytest.fixture(scope="session")
