+++ tests/galleries/test_galleries.py
@@ -1,0 +1,8 @@
+import pytest
+
+
+@pytest.mark.slow
+def test_gallery_scripts_smoke(subtests, sphinx_gallery_scripts):
+    for script in sphinx_gallery_scripts:
+        with subtests.test(script):
+            __import__(script)
