+++ .github/workflows/tests.yml
@@ -53,12 +53,41 @@ jobs:
         run: pip install -r requirements-dev.txt
 
       - name: Create environment
-        run: tox --notest
+        run: tox -e tests-integration --notest
 
       - name: Run tests
-        run: tox -- --skip-large-download
+        run: tox -e tests-integration -- --skip-large-download
 
       - name: Upload coverage
         uses: codecov/codecov-action@v1.0.12
         with:
           env_vars: OS,PYTHON
+
+  galleries:
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Set up python
+        uses: actions/setup-python@v2
+        with:
+          python-version: 3.6
+
+      - name: Upgrade pip
+        run: python -m pip install --upgrade pip
+
+      - name: Upgrade or install additional system packages
+        run: pip install --upgrade pip setuptools virtualenv wheel
+
+      - name: Checkout repository
+        uses: actions/checkout@v2
+        with:
+          fetch-depth: 0
+
+      - name: Install dev requirements
+        run: pip install -r requirements-dev.txt
+
+      - name: Create environment
+        run: tox -e tests-galleries --notest
+
+      - name: Run tests
+        run: tox -e tests-galleries
