+++ pytest.ini
@@ -5,5 +5,4 @@ markers =
   slow
   slow_if_cuda_not_available
   flaky
-testpaths = tests/
 addopts = -ra

[…]
