### 2020-08-18 aefcd4f
remove typing_extensions from install requirements
