+++ setup.cfg
@@ -41,7 +41,6 @@ install_requires =
     pillow
     numpy
     requests
-    typing_extensions
 
 [options.packages.find]
 exclude =
