### 2020-08-18 1049077
remove dependency on requests (#384)
