### 2020-08-18 45b63a7
remove typing_extensions from install requirements (#385)
