+++ setup.cfg
@@ -40,7 +40,6 @@ install_requires =
     torchvision>=0.7.0
     pillow
     numpy
-    typing_extensions
 
 [options.packages.find]
 exclude =
