### 2020-08-18 2ed285c
add tests for torch.hub
