+++ tox.ini
@@ -47,6 +47,17 @@ commands =
     tests/galleries \
     {posargs}
 
+[testenv:tests-hub]
+passenv =
+  GITHUB_ACTIONS
+  GITHUB_CONTEXT
+deps = {[tests-common]deps}
+force_cpu = True
+commands =
+  {[tests-common]commands} \
+    tests/hub \
+    {posargs}
+
 [testenv:lint-style]
 whitelist_externals =
   pre-commit
