+++ .github/workflows/tests.yml
@@ -91,3 +91,42 @@ jobs:
 
       - name: Run tests
         run: tox -e tests-galleries
+
+  hub:
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Set up python
+        uses: actions/setup-python@v2
+        with:
+          python-version: 3.6
+
+      - name: Upgrade pip
+        run: python -m pip install --upgrade pip
+
+      - name: Upgrade and install additional system packages
+        run: pip install --upgrade setuptools virtualenv wheel
+
+      - name: Checkout repository
+        uses: actions/checkout@v2
+        with:
+          fetch-depth: 0
+
+      - name: Install dev requirements
+        run: pip install -r requirements-dev.txt
+
+      - name: Create environment
+        run: tox -e tests-hub --notest
+
+      - name: Load torch.hub checkpoints from cache
+        env:
+          TORCH_HUB_CHECKPOINTS: "$(python -c 'from torch import hub; print(hub.get_dir())')/checkpoints"
+        uses: actions/cache@v2
+        with:
+          path: TORCH_HUB_CHECKPOINTS
+          key: torch-hub-checkpoints-${{ hashFiles('hubconf.py') }}
+
+      - name: Run tests
+        env:
+          GITHUB_CONTEXT: ${{ toJson(github) }}
+        run: tox -e tests-hub
