### 2020-08-18 77edda7
add tests for torch.hub (#386)
* add tests for torch.hub

* add workflow
