### 2020-08-18 f94586f
make hub models a drop in replacement for torchvision
