### 2020-08-18 eeb1292
upgrade test for vgg in hub
