### 2020-08-18 ffbbfb9
expose all vgg archs in hubconf
