+++ hubconf.py
@@ -2,8 +2,6 @@ dependencies = ["pystiche"]
 
 from pystiche.enc.models.vgg import MODELS
 
-for arch in ("vgg16", "vgg19"):
-    globals()[arch] = MODELS[arch]
+globals().update(MODELS)
 
-del arch
 del MODELS
