### 2020-08-18 c1b830e
make hub models a drop in replacement for torchvision (#387)
* make hub models a drop in replacement for torchvision

* lint

* upgrade test for vgg in hub

* add manual github overwrite via env var

* expose all vgg archs in hubconf

* fix

* trigger CI to test caching

* Revert "trigger CI to test caching"

This reverts commit 8cd315e4be5bd90dad406123f4bbd21d1c9bb1a2.
