+++ pystiche/enc/models/vgg.py
@@ -1,30 +1,15 @@
-from typing import Any, Dict, List, Tuple
+import re
+from typing import Any, Callable, Dict, List, Tuple, cast
 
 import torchvision
-from torch import nn
+from torch import hub, nn
 from torch.utils import model_zoo
 from torchvision.models import VGG
 
+from pystiche.misc import verify_str_arg
+
 from ..multi_layer_encoder import MultiLayerEncoder
 from ..preprocessing import get_preprocessor
-
-MODELS = {
-    name: vgg_net
-    for name, vgg_net in torchvision.models.vgg.__dict__.items()
-    if name.startswith("vgg") and callable(vgg_net)
-}
-
-MODEL_URLS = {
-    ("torch", arch): url for arch, url in torchvision.models.vgg.model_urls.items()
-}
-# The caffe weights were created by Karen Simonyan and Andrew Zisserman. See
-# https://download.pystiche.org/models/LICENSE for details.
-MODEL_URLS.update(
-    {
-        ("caffe", "vgg16",): "https://download.pystiche.org/models/vgg16-781be684.pth",
-        ("caffe", "vgg19",): "https://download.pystiche.org/models/vgg19-74e45263.pth",
-    }
-)
 
 __all__ = [
     "VGGMultiLayerEncoder",
@@ -37,6 +22,99 @@ __all__ = [
     "vgg19_multi_layer_encoder",
     "vgg19_bn_multi_layer_encoder",
 ]
+
+TORCH_MODEL_URLS = torchvision.models.vgg.model_urls
+ARCHS = tuple(TORCH_MODEL_URLS.keys())
+
+MODEL_URLS = {(arch, "torch"): TORCH_MODEL_URLS[arch] for arch in ARCHS}
+MODEL_URLS.update(
+    {
+        ("vgg16", "caffe"): "https://download.pystiche.org/models/vgg16-781be684.pth",
+        ("vgg19", "caffe"): "https://download.pystiche.org/models/vgg19-74e45263.pth",
+    }
+)
+
+
+NUM_LAYERS_TO_CONFIGURATION = {
+    11: "A",
+    13: "B",
+    16: "D",
+    19: "E",
+}
+
+
+def _make_vgg_docstring(num_layers: int, batch_norm: bool) -> str:
+    description = (
+        f"VGG{num_layers} model "
+        f"(configuration '{NUM_LAYERS_TO_CONFIGURATION[num_layers]}') "
+        f"from :cite:`SZ2014`"
+        f"{' with batch normalization' if batch_norm else ''}.\n"
+    )
+    args = r"""Args:
+    pretrained: If ``True``, loads weights from training on
+        :class:`~torchvision.models.ImageNet`. Defaults to ``False``.
+    framework: Framework that was used to train the model. Can be one of
+        ``"torch"`` (default) or ``"caffe"``.
+        .. note::
+
+            The weights for ``"caffe"`` were generated by Karen Simonyan and
+            Andrew Zisserman. See https://download.pystiche.org/models/LICENSE for
+            details.
+    progress: If ``True``, displays a progress bar to STDERR during download.
+        Defaults to ``True``.
+    num_classes: Size of the output layer. Defaults to ``1000``.
+        .. note::
+
+            Pretrained weights are only available for ``num_classes == 1000``.
+    """
+
+    return "\n".join((description, args))
+
+
+ARCH_PATTERN = re.compile(r"vgg(?P<num_layers>(11|13|16|19)+)(?P<batch_norm>_bn)?")
+
+
+def _vgg_loader(arch: str) -> Callable[..., torchvision.models.VGG]:
+    match = ARCH_PATTERN.match(arch)
+    if match is None:
+        raise ValueError(
[…]
