+++ tox.ini
@@ -49,6 +49,7 @@ commands =
 
 [testenv:tests-hub]
 passenv =
+  PYSTICHE_HUB_GITHUB
   GITHUB_ACTIONS
   GITHUB_CONTEXT
 deps = {[tests-common]deps}
