### 2020-08-18 ab8fdf4
make package importable without installation
