### 2020-08-18 66084ed
make package importable without installation (#388)
