+++ pystiche/__init__.py
@@ -1,4 +1,7 @@
-from ._version import version as __version__  # type: ignore[import]
+try:
+    from ._version import version as __version__  # type: ignore[import]
+except ImportError:
+    __version__ = "UNKNOWN"
 
 from .core import *
 
