### 2020-08-18 ea506da
mock weight loading to test models in CI
