### 2020-08-18 6f39808
add tests for pystiche.enc.models repr
