### 2020-08-18 09dbc76
remove false markers
