### 2020-08-18 e328316
maybe reduce memory load
