### 2020-08-19 950b848
add tests for pystiche.enc.models repr (#391)
* add tests for pystiche.enc.models repr

* fix fixture

* remove false markers

* maybe reduce memory load

* fix memory leak
