### 2020-08-19 f95ff44
refactor pystiche.enc.models
