+++ pystiche/enc/models/vgg.py
@@ -1,16 +1,16 @@
 import re
-from typing import Any, Callable, Dict, List, Tuple, cast
+import warnings
+from typing import Any, Callable, Dict, List, Optional, Tuple, cast
 
 import torchvision
 from torch import hub, nn
-from torch.utils import model_zoo
-from torchvision.models import VGG
 
-from pystiche.misc import verify_str_arg
+from pystiche.misc import build_deprecation_message
 
-from ..multi_layer_encoder import MultiLayerEncoder
-from ..preprocessing import get_preprocessor
+from .utils import ModelMultiLayerEncoder
+from .utils import select_url as _select_url
 
+# the functions are only defined at import
 __all__ = [
     "VGGMultiLayerEncoder",
     "vgg11_multi_layer_encoder",
@@ -23,17 +23,6 @@ __all__ = [
     "vgg19_bn_multi_layer_encoder",
 ]
 
-TORCH_MODEL_URLS = torchvision.models.vgg.model_urls
-ARCHS = tuple(TORCH_MODEL_URLS.keys())
-
-MODEL_URLS = {(arch, "torch"): TORCH_MODEL_URLS[arch] for arch in ARCHS}
-MODEL_URLS.update(
-    {
-        ("vgg16", "caffe"): "https://download.pystiche.org/models/vgg16-781be684.pth",
-        ("vgg19", "caffe"): "https://download.pystiche.org/models/vgg19-74e45263.pth",
-    }
-)
-
 
 NUM_LAYERS_TO_CONFIGURATION = {
     11: "A",
@@ -42,14 +31,44 @@ NUM_LAYERS_TO_CONFIGURATION = {
     19: "E",
 }
 
+ARCH_PATTERN = re.compile(r"^vgg(?P<num_layers>(11|13|16|19)+)(?P<batch_norm>_bn)?$")
 
-def _make_vgg_docstring(num_layers: int, batch_norm: bool) -> str:
-    description = (
-        f"VGG{num_layers} model "
-        f"(configuration '{NUM_LAYERS_TO_CONFIGURATION[num_layers]}') "
-        f"from :cite:`SZ2014`"
-        f"{' with batch normalization' if batch_norm else ''}.\n"
+
+def _parse_arch(arch: str) -> Tuple[int, bool]:
+    match = ARCH_PATTERN.match(arch)
+    if match is None:
+        raise ValueError(
+            f"Unknown arch '{arch}'. It has to match 'vgg_(11|13|16|19)(_bn)?'"
+        )
+
+    num_layers = int(match.group("num_layers"))
+    batch_norm = match.group("batch_norm") is not None
+
+    return num_layers, batch_norm
+
+
+def _make_description(arch: str, multi_layer_encoder: bool) -> str:
+    num_layers, batch_norm = _parse_arch(arch)
+
+    if multi_layer_encoder:
+        short = (
+            f"Multi-layer encoder based on :class:`~torchvision.models.VGG` "
+            f"{num_layers}{' with batch normalization' if batch_norm else ''}."
+        )
+    else:
+        short = f":class:`~torchvision.models.VGG` {num_layers} model"
+
+    long = (
+        f"    The :class:`~torchvision.models.VGG` architecture was introduced by "
+        f"Krizhevsky, Sutskever, and Hinton in :cite:`KSH2012`. VGG{num_layers} "
+        f"corresponds to configuration ``{NUM_LAYERS_TO_CONFIGURATION[num_layers]}`` "
+        f"in the paper."
     )
+    return "\n".join((short, "", long))
+
+
+def _make_vgg_docstring(arch: str) -> str:
+    description = _make_description(arch, multi_layer_encoder=False)
     args = r"""Args:
[…]
