+++ pystiche/enc/models/__init__.py
@@ -1,2 +1,3 @@
 from .alexnet import *
+from .utils import *
 from .vgg import *
