### 2020-08-20 354221a
move download logic into utils
