+++ pystiche/enc/models/utils.py
@@ -2,13 +2,13 @@ from abc import abstractmethod
 from typing import Any, Callable, Dict, List, Optional, Tuple, TypeVar, cast
 
 import torch
-from torch import nn
+from torch import hub, nn
 from torch.nn.modules.module import _IncompatibleKeys
 
 from ..multi_layer_encoder import MultiLayerEncoder
 from ..preprocessing import get_preprocessor
 
-__all__ = ["ModelMultiLayerEncoder"]
+__all__ = ["ModelMultiLayerEncoder", "select_url"]
 
 T = TypeVar("T")
 
@@ -22,11 +22,7 @@ def select_url(
     try:
         return urls[key]
     except KeyError:
-        msg = (
-            f"No pre-trained weights are available for the parameter configuration\n\n"
-            f"{format(key)}"
-        )
-        raise RuntimeError(msg)
+        raise RuntimeError(f"No URL is available for\n\n{format(key)}")
 
 
 class ModelMultiLayerEncoder(MultiLayerEncoder):
@@ -59,23 +55,34 @@ class ModelMultiLayerEncoder(MultiLayerEncoder):
         self.internal_preprocessing = internal_preprocessing
         self.allow_inplace = allow_inplace
 
-        modules, self._state_dict_key_map = self.collect_modules(
-            pretrained, framework, allow_inplace
-        )
+        modules, self._state_dict_key_map = self.collect_modules(allow_inplace)
         if internal_preprocessing:
             modules.insert(0, ("preprocessing", get_preprocessor(framework)))
 
         super().__init__(modules)
 
+        if pretrained:
+            self.load_state_dict_from_url(framework)
+
+    @abstractmethod
+    def state_dict_url(self, framework: str) -> str:
+        r"""Select URL of a downloadable ``state_dict``.
+
+        Args:
+            framework: Name of the framework that was used to train the weights.
+
+        Raises:
+            RuntimeError: If no ``state_dict`` is available.
+        """
+        pass
+
     @abstractmethod
     def collect_modules(
-        self, pretrained: bool, framework: str, inplace: bool
+        self, inplace: bool
     ) -> Tuple[List[Tuple[str, nn.Module]], Dict[str, str]]:
         r"""Collect modules of a base model with more descriptive names.
 
         Args:
-            pretrained: If ``True``, should load builtin weights.
-            framework: Name of the framework that was used to train the builtin weights.
             inplace: If ``True``, when possible, modules should use inplace operations.
 
         Returns:
@@ -106,10 +113,6 @@ class ModelMultiLayerEncoder(MultiLayerEncoder):
     ) -> _IncompatibleKeys:
         r"""Loads parameters and buffers from the ``state_dict``.
 
-        .. seealso::
-
-            :meth:`torch.nn.Module.load_state_dict`
-
         Args:
             state_dict: State dictionary.
             strict: Enforce matching keys in ``state_dict`` and the internal states.
@@ -126,6 +129,10 @@ class ModelMultiLayerEncoder(MultiLayerEncoder):
 
         Returns:
             Named tuple with ``missing_keys`` and ``unexpected_keys`` fields.
+
+        .. seealso::
+
+            :meth:`torch.nn.Module.load_state_dict`
         """
         if map_names:
             state_dict, unexpected_keys = self._map_state_dict_keys(state_dict)
@@ -141,6 +148,39 @@ class ModelMultiLayerEncoder(MultiLayerEncoder):
         self.framework = framework
 
         return keys
+
+    def load_state_dict_from_url(
+        self,
+        framework: str,
+        strict: bool = True,
+        map_names: bool = True,
+        check_hash: bool = True,
+        **kwargs: Any,
+    ) -> None:
+        r"""Downloads and loads parameters and buffers trained with ``framework``.
+
+        Args:
[…]
