### 2020-08-20 97cb388
fix sphinx warning for vgg
