+++ tests/integration/enc/test_models.py
@@ -1,11 +1,19 @@
+import functools
 import itertools
+from collections import OrderedDict
+from copy import copy
+from urllib.parse import urljoin
 
 import pytest
+import pytorch_testing_utils as ptu
 
 import torch
+from torch import nn
 
 import pystiche
 from pystiche import enc
+
+from tests import asserts, mocks
 
 
 @pytest.mark.large_download
@@ -96,3 +104,215 @@ def test_vgg_multi_layer_encoder_smoke(
 
             with subtests.test("repr"):
                 assert isinstance(repr(multi_layer_encoder), str)
+
+
+def test_select_url():
+    key = "available"
+    url = "https://pystiche.org"
+
+    assert enc.select_url({key: url}, key) == url
+
+
+def test_select_url_not_available():
+    key = "not_available"
+    urls = {}
+
+    with pytest.raises(RuntimeError):
+        enc.select_url(urls, key)
+
+
+def test_select_url_custom_format():
+    key = "not_available"
+    urls = {}
+
+    def format(key):
+        return f"custom format for '{key}'"
+
+    expected = format(key)
+
+    with pytest.raises(RuntimeError) as exc_info:
+        enc.select_url(urls, key, format=format)
+
+        actual = exc_info.value.args[0]
+        assert actual.endswith(expected)
+
+
+@pytest.fixture
+def frameworks():
+    return ("torch", "caffe")
+
+
+@pytest.fixture
+def multi_layer_encoder_urls(frameworks):
+    base = "https://download.pystiche.org/models/"
+
+    return {
+        framework: urljoin(base, f"{framework}-01234567.pth")
+        for framework in frameworks
+    }
+
+
+@pytest.fixture
+def multi_layer_encoder_modules():
+    return [("conv", nn.Conv2d(3, 3, 1)), ("relu", nn.ReLU())]
+
+
+@pytest.fixture
+def multi_layer_encoder_state_dict(multi_layer_encoder_modules):
+    return OrderedDict(
+        (f"{name}.{key}", tensor)
+        for name, module in multi_layer_encoder_modules
+        for key, tensor in module.state_dict().items()
+    )
+
+
+@pytest.fixture
+def other_multi_layer_encoder_state_dict(multi_layer_encoder_state_dict):
+    return OrderedDict(
+        (id, torch.zeros_like(tensor))
+        for id, tensor in multi_layer_encoder_state_dict.items()
+    )
+
+
+@pytest.fixture
+def multi_layer_encoder_state_dict_key_map(multi_layer_encoder_modules):
+    return {
+        f"{idx}.{key}": f"{name}.{key}"
+        for idx, (name, module) in enumerate(multi_layer_encoder_modules)
+        for key in module.state_dict().keys()
+    }
+
+
+@pytest.fixture
+def multi_layer_encoder_cls(
+    mocker,
+    multi_layer_encoder_urls,
+    multi_layer_encoder_modules,
+    multi_layer_encoder_state_dict_key_map,
+):
+    class MockModelMultiLayerEncoder(enc.ModelMultiLayerEncoder):
+        def state_dict_url(self, framework: str) -> str:
+            return enc.select_url(multi_layer_encoder_urls, framework)
+
+        def collect_modules(self, inplace: bool):
+            return (
+                copy(multi_layer_encoder_modules),
+                multi_layer_encoder_state_dict_key_map,
+            )
+
+    mocker.patch(
+        mocks.make_mock_target(
+            "enc", "models", "utils", "hub", "load_state_dict_from_url"
+        ),
+        return_value=MockModelMultiLayerEncoder(pretrained=False).state_dict(),
+    )
+
+    return MockModelMultiLayerEncoder
+
+
+def test_ModelMultiLayerEncoder_internal_preprocessing(
[…]
