### 2020-08-20 3fd3504
refactor pystiche.enc.models (#392)
* refactor pystiche.enc.models

* move download logic into utils

* fix sphinx warning for vgg

* fix tests

* add tests
