+++ pystiche/enc/models/alexnet.py
@@ -1,62 +1,72 @@
-from typing import Any, Dict, List, Tuple
+import warnings
+from typing import Any, Dict, List, Optional, Tuple
 
 from torch import nn
-from torch.utils import model_zoo
-from torchvision.models import AlexNet, alexnet
+from torchvision.models import alexnet
+from torchvision.models.alexnet import model_urls as TORCH_MODEL_URLS
 
-from ..multi_layer_encoder import MultiLayerEncoder
-from ..preprocessing import get_preprocessor
+from pystiche.misc import build_deprecation_message
 
-MODEL_URLS = {"torch": "https://download.pytorch.org/models/alexnet-owt-4df8aa71.pth"}
-
+from .utils import ModelMultiLayerEncoder, select_url
 
 __all__ = ["AlexNetMultiLayerEncoder", "alexnet_multi_layer_encoder"]
 
 
-class AlexNetMultiLayerEncoder(MultiLayerEncoder):
-    r"""Multi-layer encoder based on the AlexNet architecture that was introduced by
+MODEL_URLS = {"torch": TORCH_MODEL_URLS["alexnet"]}
+
+
+def _make_description() -> str:
+    return r"""Multi-layer encoder based on :class:`~torchvision.models.AlexNet`.
+
+    The :class:`~torchvision.models.AlexNet` architecture was introduced by
     Krizhevsky, Sutskever, and Hinton in :cite:`KSH2012`.
-
-    Args:
-        internal_preprocessing: If ``True``, adds a preprocessing layer for the
-            selected ``weights`` as first layer. Defaults to ``"True"``.
-        allow_inplace: If ``True``, allows inplace operations in the
-            :class:`torch.nn.Relu` layers to reduce the memory requirement during the
-            forward pass. Defaults to ``False``.
-
-            .. warning::
-                After performing an inplace operation in a :class:`torch.nn.Relu` layer
-                the encoding of the previous :class:`torch.nn.Conv2d` layer is no
-                longer accessible. Only set this to ``True`` if you are sure that you
-                do **not** need the encodings of the :class:`torch.nn.Conv2d` layers.
     """
 
-    def __init__(
-        self,
-        weights: str = "torch",
-        internal_preprocessing: bool = True,
-        allow_inplace: bool = False,
-    ) -> None:
-        self.weights = weights
-        self.internal_preprocessing = internal_preprocessing
-        self.allow_inplace = allow_inplace
-        super().__init__(self._collect_modules())
 
-    def _collect_modules(self) -> List[Tuple[str, nn.Module]]:
-        base_model = alexnet(pretrained=False)
-        self._load_weights(base_model)
+def _make_docstring(body: str) -> str:
+    return f"{_make_description()}\n{body}"
 
-        model = base_model.features
+
+class AlexNetMultiLayerEncoder(ModelMultiLayerEncoder):
+    __doc__ = _make_docstring(
+        r"""    Args:
+        pretrained: If ``True``, loads builtin weights. Defaults to ``True``.
+        framework: Name of the framework that was used to train the builtin weights.
+            Defaults to ``"torch"``.
+        kwargs: Optional arguments of :class:`ModelMultiLayerEncoder` .
+
+    Raises:
+        RuntimeError: If ``pretrained`` and no weights are available for the
+        ``framework``.
+        """
+    )
+
+    def __init__(self, weights: Optional[str] = None, **kwargs: Any) -> None:
+        if weights is not None:
+            msg = build_deprecation_message(
+                "The parameter weights", "0.6.0", info="It was renamed to framework"
+            )
+            warnings.warn(msg, UserWarning)
+            kwargs["framework"] = weights
+
+        super().__init__(**kwargs)
+
[…]

[…]
