### 2020-08-20 428c85a
add remaining tests for alexnet
