+++ tests/integration/enc/models/conftest.py
@@ -1,0 +1,6 @@
+import pytest
+
+
+@pytest.fixture(scope="package")
+def frameworks():
+    return ("torch", "caffe")
