### 2020-08-20 3cef669
add remaining tests for VGG
