### 2020-08-20 aace2d8
add remaining tests for pystiche.enc.models (#393)
* split model tests in separate modules

* add remaining tests for alexnet

* add remaining tests for VGG
