### 2020-08-20 d8c6c59
add hubconf to test triggers
