### 2020-08-20 dbe339b
fix path discovery
