### 2020-08-20 0373969
only cache vgg 16 and 19
