### 2020-08-21 7890fb8
fix annotations for vgg multi-layer encoders
