### 2020-08-25 fbd78b9
adapt hub tests
