### 2020-08-25 a909afb
Fix checkpoints caching for hub tests (#394)
* try fix

* add hubconf to test triggers

* add debug

* try trigger caching

* fix path discovery

* try fix

* fix

* only cache vgg 16 and 19

* fix workflow

* adapt hub tests

* re-trigger CI
