### 2020-08-25 08365c3
revert dynamic creation of vgg multi_layer_encoder s (#397)
