### 2020-08-28 #25 @pmeier
@NickleDave @lwasser

Could we get an update on this? Its been almost a month. If I can help in any way, please let me know.

My grant is running until the end of the year and I need this published (in JOSS) before that. I'm aiming to have this published in late November or early December since I have some other duties at the end of the year. This leaves us with approximately 3 months from now **including** the JOSS process. Is this still reasonable together with pyOpenSci? If not, I need to take this to JOSS directly although I really want the pyOpenSci review.
