### 2020-08-29 #25 @NickleDave
Hi @pmeier I am sorry I have not gotten back to you, totally my fault.

Unfortunately I have not heard back from multiple invited reviewers.

I think part of it might also be that the package targets a relatively new and developing research area, and it's hard to find someone that has the right mix of domain expertise and OSS development experience.

I really appreciate that you want the scientific rigor added by our review.
Let's go ahead with just @soumith and @edgarriba as reviewers.
I'm sure their valuable input will help speed up review at JOSS as well.

@soumith and @edgarriba can you commit to finishing review two weeks from now, by **Saturday September 12th**, so we can help Philip get this to JOSS?
I'm sure you are also both busy, and I don't mean to make it your problem that I have not been more on top of this review.
But I think we all have some idea of what it's like to be a PhD student whose grant is running out.

It would be great if you can do a quick review that mainly focuses on the abstractions the library provides for this research.
JOSS reviewers will be able to address questions of software engineering, so you do not need to spend as much time on those (even though I know you could). But you can give big picture feedback on the goals of the library and how it achieves those, given your knowledge of deep learning and computer vision.

Please @soumith @edgarriba let me know if that deadline of Saturday September 12th works for you, and if not, when you expect you will be able to complete review.
That will help @pmeier decide how to move forward.
