### 2020-08-29 #25 @pmeier
@NickleDave I appreciate the change of course. I don't want to interfere with the process, just wanted to say that we don't need to rush it that much. I think it should be enough if I get the reviews until the end of September.
