### 2020-08-29 #25 @NickleDave
ok, thank you for that additional information

if we start today (ignoring that I have made put this review very behind) then the standard three weeks deadline would put us at September 19th.

I am going to say September 19th and make sure I have that in my calendar so I stay on top of this, unless @edgarriba and @soumith absolutely cannot make that deadline
