### 2020-08-31 #25 @lwasser
hey all!! just a note that once the review is complete here - the JOSS process is fast as they fully accept our review. So the key will be getting the reviewers on board with the Sept deadline and then addressing any feedback promptly. Thank you for your time. Just a note i've been struggling a bit with keeping up given COVID, beginning of the semester and other issues happening in the world right now. it would be great to find funding for this project so someone can work on it full time! i just haven't had time to devote to it.
