### 2020-09-03 #25 @NickleDave
🙏 🙏 🙏 thank you @soumith !!!

@edgarriba can you please confirm whether you can make that deadline when you get a chance?
