+++ tox.ini
@@ -76,7 +76,7 @@ commands =
 
 [sphinx-gallery]
 deps =
-  sphinx-gallery>=0.7.0
+  sphinx-gallery>=0.8
   # Additional sphinx-gallery dependencies
   # https://sphinx-gallery.github.io/stable/index.html#install-via-pip
   matplotlib
