+++ .pre-commit-config.yaml
@@ -5,7 +5,7 @@ repos:
       - id: isort
         args: [--settings-path=pyproject.toml]
   - repo: https://github.com/psf/black
-    rev: stable
+    rev: "19.10b0"
     hooks:
       - id: black
         args: [--config=pyproject.toml]
