### 2020-09-06 a168a10
limit black version to 19.10b0 (#398)
* limit black version

* add pre-commit config as lint workflow trigger
