+++ .pre-commit-config.yaml
@@ -6,7 +6,7 @@ repos:
         args: [--settings-path=pyproject.toml, --filter-files]
         additional_dependencies: [toml]
   - repo: https://github.com/psf/black
-    rev: stable
+    rev: "19.10b0"
     hooks:
       - id: black
         args: [--config=pyproject.toml]
