### 2020-09-19 #25 @edgarriba
- this is missing for development mode.
  - not in the readme. Working examples just in the docs.
- [x] **Function Documentation:** for all user-facing functions
- [x] **Examples** for all user-facing functions
  - missing in docs usage example for some functions.
  - not linked in the README.
- [x] **Metadata** including author(s), author e-mail(s), a url, and any other relevant metadata e.g., in a `setup.py` file or elsewhere.

Readme requirements

- [x] Badges for continuous integration and test coverage, a repostatus.org badge, and any other badges. If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be more wide than high. (Note that the badge for pyOpenSci peer-review will be provided upon acceptance.)
- [ ] Short description of goals of package, with descriptive links to all vignettes (rendered, i.e. readable, cf the documentation website section) unless the package is small and there’s only one vignette repeating the README.
- [ ] Any additional setup required (authentication tokens, etc)
  - not sure if it's applicable here.
- [ ] Brief demonstration usage
  - missing
- [x] Direction to more detailed documentation (e.g. your documentation files or website).
- [x] If applicable, how the package compares to other similar packages and/or how it relates to other packages
  - missing

Package structure should follow general community best-practices. In general please consider:

- [x] The documentation is easy to find and understand
  - missing usage examples in the docs

  - ERROR tests/integration/enc/models/test_vgg.py
- [x] **Continuous Integration:** Has continuous integration, such as Travis CI, AppVeyor, CircleCI, and/or others.
  - provided link for packaging guidelines doesn't work.

#### For packages co-submitting to JOSS

---

The package looks good overall with a good organization and using updated tools to keep high quality standards in terms of software engineering. The repository seems to have a simple and proper structure for the specific goals of the project. The documentation seems quite complete including few but clear working examples in addition to a good and complete package reference.

The only weakness I see in the package is the development support. I list below few comments and suggestions on that:

- Apart from pip install I miss short git clone instructions in the README. Add short development instructions in the README about how to git clone the project since probably users landing to github are more development oriented. Might be the case that not everyone is familiar with git.
- I miss a link to the CONTRIBUTING doc in the README.
  - inside the contributing guides I suggest to include also some initial explanation about setting up the development environment and make a fresh check for the current branch.
  - Missing some sort of script to create a virtual environment for development to avoid installing the dependencies in the main system. What if you don't even have pip, or tox ? :)
  - Needed to install all the dependencies and tools by hand to get a success in tests.
- Docs are okay, just missing very few documentation (like in the image module). I would expect to see a small usage example in the most relevant classes.
- I see few usage for type or shape checking and raising informative errors in some of the functionalities.
