### 2020-09-20 #25 @pmeier
Hey @edgarriba, thanks a lot for your review. I'm only going to comment on things I think need discussion, i.e. I'll omit comments I agree with.

---

> - [ ] Brief demonstration usage
>  missing

You are of course correct here, but I need to challenge the intent. Thus, this comment is more directed at @NickleDave. `pystiche` is only valuable because it provides all pieces of an NST that work together. The individual parts are probably not helpful to anyone. Thus, I only have usage examples that show the complete workflow. The minimal example I came up with is [this](https://pystiche.readthedocs.io/en/latest/galleries/examples/beginner/example_nst_with_pystiche.html#sphx-glr-galleries-examples-beginner-example-nst-with-pystiche-py). IMO this is far too long to be included in a README and thus I left it out. Is the "brief usage demonstration" in the README a hard requirement?

---

> - [ ] Citation information
>    missing

Correct again, but I'm not sure how to handle this properly. `pystiche` will be published in JOSS and of course I will add the citation information after this is done. In the mean time the only thing I could include is a preliminary citation, which I would rather not do. If possible, I would have `pystiche` only citable **after** it is peer-reviewed. @NickleDave what is the way to go here?

---

> Missing some sort of script to create a virtual environment for development to avoid installing the dependencies in the main system. What if you don't even have pip, or tox ? :)

I think I can require some basic prerequisite before anyone starts developing with `python` and I think having `pip` is one of them. This is not me judging someone just starting out, since everyone starts somewhere. I just don't think if you don't know even how to do this very basic steps, your first move should be trying to contribute to an (somewhat complex) project.

I'm torn on the virtual environment. On one hand I use one for the reason you suggested, but on the other hand its only very minimal [requirements](https://github.com/pmeier/pystiche/blob/master/requirements-dev.txt): `pre-commit` and `tox`. For everything else `tox` handles the virtual environments.

---

> Needed to install all the dependencies and tools by hand to get a success in tests.

Could you share your process with me? If you correctly use `tox` (as explained in the [contributing guidelines](https://pystiche.readthedocs.io/en/latest/contributing.html)) you do not need to install anything by hand other than the minimal development requirements.

---

> - [ ] Automated tests: Tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.
>    ERROR tests/integration/enc/models/test_vgg.py

Again, could you share your process and the error message with me? Everything is passing locally and in CI for me.

---

> I see few usage for type or shape checking and raising informative errors in some of the functionalities.

Could be more specific. Where do you see an opening for better error messages?

---

> Docs are okay, just missing very few documentation (like in the image module). I would expect to see a small usage example in the most relevant classes.

I agree with the first part. Could you give me an example for the second part?
