### 2020-09-21 pmeier/pystiche#399 @pmeier
Improve README and CONTRIBUTING
As suggested in [the review](https://github.com/pyOpenSci/software-review/issues/25#issuecomment-695366229) by @edgarriba:

- Add the package goals to the README
- Add installation instructions to the README
- Add contributing guidelines to the READMe and add more information about the environment setup / `git` workflow
