+++ docs/source/index.rst
@@ -1,35 +1,14 @@
 Welcome to ``pystiche`` 's documentation!
 =========================================
 
-``pystiche`` is a framework for
-`Neural Style Transfer (NST) <https://github.com/ycjing/Neural-Style-Transfer-Papers>`_
-built upon `PyTorch <https://pytorch.org>`_. An NST allows enables a fully automatic
-merge of arbitrary content and artistic style. The name of the project is a pun on
-*pastiche* `meaning <https://en.wikipedia.org/wiki/Pastiche>`_:
-
-    A pastiche is a work of visual art [...] that imitates the style or character of
-    the work of one or more other artists. Unlike parody, pastiche celebrates, rather
-    than mocks, the work it imitates.
-
-``pystiche`` has similar goals as Deep Learning (DL) frameworks such as PyTorch:
-
-1. **Accessibility**
-    Starting off with NST can be quite overwhelming due to the sheer amount of
-    techniques one has to know and be able to deploy. ``pystiche`` aims to provide an
-    easy-to-use interface that reduces the necessary prior knowledge about NST and DL
-    to a minimum.
-2. **Reproducibility**
-    Implementing NST from scratch is not only inconvenient but also error-prone.
-    ``pystiche`` aims to provide reusable tools that let developers focus on their
-    ideas rather than worrying about bugs in everything around it.
+.. include:: general.rst
 
 .. toctree::
   :maxdepth: 1
   :caption: Contents
 
-  Installation <installation>
-  Gist <gist>
-  Examples <galleries/examples/index>
-  Contributing <contributing>
+  Getting Started <getting_started/index>
+  Gist <gist/index>
+  Usage Examples <galleries/examples/index>
   Package Reference <api/index>
   Literature Reference <literature>
