### 2020-09-21 9c8e286
add github workflow example to contributing guidelines
