### 2020-09-21 f77b4a6
Improve README and CONTRIBUTING (#400)
* restructure documentation files

* add github workflow example to contributing guidelines

* update readme

* remove include from README

* add note for PR opening
