+++ docs/source/getting_started/index.rst
@@ -1,0 +1,8 @@
+Getting started
+===============
+
+.. toctree::
+  :maxdepth: 1
+
+  Installation <installation>
+  Contributing <contributing>
