+++ CONTRIBUTING.rst
@@ -1,27 +1,145 @@
-Contributing guide lines
-========================
+Contributing
+============
 
-We appreciate all contributions. If you are planning to contribute bug-fixes or
-documentation improvements, please open a
+First and foremost: Thank you for your interest in ``pystiche`` s development! We
+appreciate all contributions be it code or something else.
+
+Guide lines
+-----------
+
+``pystiche`` uses the `GitHub workflow <https://guides.github.com/introduction/flow/>`_
+. Below is small guide how to make your first contribution.
+
+.. note::
+
+  The following guide assumes that `git <https://git-scm.com/>`_,
+  `python <https://www.python.org/>`_, and `pip <https://pypi.org/project/pip/>`_ ,
+  are available on your system. If that is not the case, follow the official
+  installation instructions.
+
+.. note::
+
+  ``pystiche`` officially supports Python ``3.6``, ``3.7``, and ``3.8``. To ensure
+  backwards compatibility, the development should happen on the minimum Python
+  version, i. e. ``3.6``.
+
+1. Fork ``pystiche`` on GitHub
+
+  Navigate to `pmeier/pystiche <https://github.com/pmeier/pystiche>`_ on GitHub and
+  click the **Fork** button in the top right corner.
+
+2. Clone your fork to your local file system
+
+  Use ``git clone`` to get a local copy of ``pystiche`` s repository that you can work
+  on:
+
+  .. code-block:: sh
+
+    $ PYSTICHE_ROOT="pystiche"
+    $ git clone "https://github.com/pmeier/pystiche.git" $PYSTICHE_ROOT
+
+3. Setup your development environment
+
+  .. code-block:: sh
+
+    $ cd $PYSTICHE_ROOT
+    $ virtualenv .venv --prompt="(pystiche) "
+    $ source .venv/bin/activate
+    $ pip install -r requirements-dev.txt
+    $ pre-commit install
+
+  .. note::
+
+    While ``pystiche`` s development requirements are fairly lightweight, it is still
+    recommended to install them in a virtual environment rather than system wide. If you
+    do not have ``virtualenv`` installed, you can do so by running
+    ``pip install --user virtualenv``.
+
+4. Create a branch for local development
+
+  Use ``git checkout`` to create local branch with a descriptive name:
+
+  .. code-block::
+
+    $ PYSTICHE_BRANCH="my-awesome-feature-or-bug-fix"
+    $ git checkout -b $PYSTICHE_BRANCH
+
+  Now make your changes. Happy Coding!
+
+5. Use ``tox`` to run various checks
+
+  .. code-block::
+
+    $ tox -f lint
+    $ tox
+    $ tox -f docs  # only if applicable
+
+  You can find details what the individual commands do below of this guide.
+
+6. Commit and push your changes
+
+  If all checks are passing you can commit your changes an push them to your fork:
+
+  .. code-block::
+
+    $ git add .
+    $ git commit -m "Descriptive message of the changes made"
+    $ git push -u origin $PYSTICHE_BRANCH
+
+  .. note::
+
+    For larger changes, it is good practice to split them in multiple small commits
+    rather than one large one. If you do that, make sure to run the test suite before
+    every commit. Furthermore, use ``git push`` without any parameters for consecutive
+    commits.
+
+7. Open a Pull request (PR)
+
+  1. Navigate to `pmeier/pystiche/pulls <https://github.com/pmeier/pystiche/pulls>`_ on
+     GitHub and click on the green button "New pull request".
+  2. Click on "compare across forks" below the "Compare changes" headline.
[…]
