### 2020-09-22 #25 @soumith
hey everyone, still working on the review, but I should have something by EOD on the 23rd.
