### 2020-09-24 #25 @NickleDave
Thank you @soumith !
And thank you @pmeier for your quick replies to Edgar's comments.
I was surviving a work deadline but will get you feedback tomorrow AM
