### 2020-09-24 #25 @soumith
- [?] **Function Documentation:** for all user-facing functions
  - not all but ~90% -- see review comments below
- [ ] **Examples** for all user-facing functions
  - I didn't see inline doc examples for almost any function. Is this the expectation of this review line item?
- [x] **Metadata** including author(s), author e-mail(s), a url, and any other relevant metadata e.g., in a `setup.py` file or elsewhere.

Readme requirements

- [x] Badges for continuous integration and test coverage, a repostatus.org badge, and any other badges. If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be more wide than high. (Note that the badge for pyOpenSci peer-review will be provided upon acceptance.)
- [x] Short description of goals of package, with descriptive links to all vignettes (rendered, i.e. readable, cf the documentation website section) unless the package is small and there’s only one vignette repeating the README.
- [x] Any additional setup required (authentication tokens, etc)
- [ ] Brief demonstration usage
- [x] Direction to more detailed documentation (e.g. your documentation files or website).
- [x] If applicable, how the package compares to other similar packages and/or how it relates to other packages

Package structure should follow general community best-practices. In general please consider:

- [x] The documentation is easy to find and understand
- [?] All functions have documentation and associated examples for use
  - ~90% functions have documentation, and the repository examples cover a lot of beginner to complex cases, so I consider this still sufficient

  - see my notes below, I had trouble running the example out of the box
  - no performance claims were made by software
- [?] **Automated tests:** Tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.
  - see notes below, had trouble running tests out of the box using instructions. I had to manually install the right packages to get tests working.
- [x] **Continuous Integration:** Has continuous integration, such as Travis CI, AppVeyor, CircleCI, and/or others.
- [?] **Packaging guidelines**: The package conforms to the pyOpenSci [packaging guidelines](https://www.pyopensci.org/dev_guide/authoring/overview.html).
  - the link to the packaging guidelines is broken

#### For packages co-submitting to JOSS

  - the references section is empty, that doesn't look right: https://github.com/pmeier/pystiche/blob/v0.5.1/paper.md

---

The package makes the whole suite of neural style transfer work much easier to comprehend and reliably run.
The documentation was nice, the examples were a breeze to read through. There's a lot of value in such a package and it's obvious that a lot of effort and tasteful design have been put into it's development.

Bigger points:

- The overall API documentation as well as examples are in really good shape. I don't have any real comments here. In-line examples for some of the functional operators (in pystiche.ops and pystiche.ops.functional) would have been nice.
- I reviewed v0.5.1 (latest stable release) and there seem to be teething problems with the environment I setup (a fresh conda env with python 3.6, and then following package instructions). I've documented it in detail enough that the author should be able to resolve this quickly.
  - Unfortunately, I wasn't able to run the examples locally due to an image hash issue that I've documented below.
I also had some issues with running unit tests locally, again documented below. w.r.t the unit tests, I think the `tox` environment isn't correctly configured. Additionally, some unit tests failed.

##### install issue

To pip install the package, I tried what was listed in the README, but I got this error:

```
pip install https://github.com/pmeier/pystiche@v0.5.1
Collecting https://github.com/pmeier/pystiche@v0.5.1
  ERROR: HTTP error 404 while getting https://github.com/pmeier/pystiche@v0.5.1
ERROR: Could not install requirement https://github.com/pmeier/pystiche@v0.5.1 because of HTTP error 404 Client Error: Not Found for url: https://github.com/pmeier/pystiche@v0.5.1 for URL https://github.com/pmeier/pystiche@v0.5.1
```

I instead used `pip install git+https://github.com/pmeier/pystiche@v0.5.1`

This is my `pip` version to help debugging the issue:

```
pip --version
pip 20.2.2 from /home/bluebox/miniconda3/envs/pystiche_review/lib/python3.6/site-packages/pip (python 3.6)
```

##### missing documentation

pystiche.meta, pystiche.misc.build_complex_obj_repr, pystiche.ops.functional.value_range_loss, some functions in pystiche.image, most functions in pystiche.image.transforms didn't have any documentation beyond input/return type information

##### issue with running the example

I had issue running the first example `example_nst_with_pystiche.py` downloaded from https://pystiche.readthedocs.io/en/stable/_downloads/36c76ed9ccf74ecc12b41424d1ea227a/example_nst_with_pystiche.py which was linked from the end of this page: https://pystiche.readthedocs.io/en/stable/galleries/examples/beginner/example_nst_with_pystiche.html#sphx-glr-galleries-examples-beginner-example-nst-with-pystiche-py

The exact error was:

```
  File "example_nst_with_pystiche.py", line 116, in <module>
    images.download()
  File "/home/bluebox/miniconda3/envs/pystiche_review/lib/python3.6/site-packages/pystiche/data/collections/download.py", line 200, in download
    image.download(root=root, overwrite=overwrite)
  File "/home/bluebox/miniconda3/envs/pystiche_review/lib/python3.6/site-packages/pystiche/data/collections/download.py", line 144, in download
    raise FileExistsError(msg)
FileExistsError: bird1.jpg with a different MD5 hash already exists in /home/bluebox/.cache/pystiche. If you want to overwrite it, set overwrite=True.
```

I have never installed pystiche on this machine before. To double-confirm, I removed `rm -rf /home/bluebox/.cache/pystiche` and reran the script again, but the error is the same.

To help debug:

```
python --version
Python 3.6.12 :: Anaconda, Inc.

pip list|grep torch
torch             1.6.0
torchvision       0.7.0

pip list|grep pystiche
pystiche          0.5.1
```

##### issue with running unit tests

Running tests didn't work out of the box. I was following the instructions in CONTRIBUTING.rst: https://github.com/pmeier/pystiche/blob/v0.5.1/CONTRIBUTING.rst#tests

See the log for details on the environment, commands used, exact output, etc: https://gist.github.com/soumith/70f989dbd9d9d5702cbb24143e88755c

I decided to try get these to work manually by calling `pytest .` and fixing issues.
I had to do the following:

```
pip install dill
pip install https://github.com/pmeier/pyimagetest/archive/v0.2.zip # i.e. the previous version of pyimagetest
pip install pillow-affine
pip install pytest-subtests
```

After this, `pytest .` started working, but one test failed: https://gist.github.com/soumith/690f8fdc9a224f85a0484a04f65d40ec
`pytest . --skip-slow` worked to completion
