### 2020-09-24 #25 @pmeier
Hey @soumith, thanks a lot for your review. IMO all your points are valid. Although I was unable to reproduce the error you had with the example exactly I've stumbled over a similar one. It seems something about the downloading / storing of the demo images is buggy.
