### 2020-09-24 #25 @NickleDave
@pmeier still digesting reviews, but here's replies to two of your questions above:

> Is the "brief usage demonstration" in the README a hard requirement?
I don't think we need to make anything a hard requirement.
But it will help users understand what the library does if you can provide *any* kind of code snippet + images in addition to the verbal description you have now. For example, why not have as a snippet lines 176-204 from your example, with the content, style, and result images? You can say something like "snippet from the complete that you can run [here](link)" to indicate to readers that they should not expect the snippet to run on its own.

You want it to be blindingly obvious from the README what the library does. Images of style transfer are the best way to do that. A code snippet that gives some flavor of how people create those images will help anyone understand how they will use your library to achieve NST, even if it's not a complete example. You don't want to lose someone because they had to dig through the docs.

> If possible, I would have pystiche only citable after it is peer-reviewed. @NickleDave what is the way to go here?
You already have a DOI that people can cite in your repo, in the badge section. That's what this "citation" refers to. So you already met this criterion. As you probably know, people often have a separate section at the bottom of the README saying "if you use this software please cite the DOI". I would strongly encourage you to add it, just to make it as clear as you can that people should be citing (your) software. You might have noticed that it's harder to reproduce results when people don't cite the versions they use :)

re: JOSS, from their guidelines:
> Upon successful completion of the review, authors will make a tagged release of the software, and deposit a copy of the repository with a data-archiving service such as Zenodo or figshare, get a DOI for the archive, and update the review issue thread with the version number and DOI.
So the DOI you have now will just change to the one for the new release associated with the JOSS review
