### 2020-09-24 #25 @NickleDave
@edgarriba do you feel @pmeier has addressed your comments?

If you would like to provide more specifics on some of the follow-up questions, e.g. about examples in docstrings, then that could be helpful for Philip.

But other than that I take it you feel the library can be approved as-is.
Thank you again for making time to do the review.
