### 2020-09-24 #25 @NickleDave
@pmeier I see that @soumith made similar comments about a lack of inline examples, with suggestions for which functions would best benefit from them.

I saw you said that all the points in the review are valid, and you might be planning to get back with a detailed reply, but given that both reviewers made comments, please do clarify whether you plan to provide those.
