### 2020-09-24 #25 @pmeier
@NickleDave @edgarriba @soumith

Here is my plan to address the review comments:

- [ ] add citation information to the README
- [ ] add a small usage snippet to the README
- [ ] fix the installation instructions and environment setup (partially done in pmeier/pystiche#400)
- [ ] fix the download and storage of the demo images
- [ ] add inline examples for `pystiche.ops` and `pystiche.ops.functional` as they form the core parts of an NST

Furthermore, if not required to pass the review, I'll postpone adding more docstrings / examples to functions and classes that I suspect will not be used by the user. That includes the "undocumented" parts of `pystiche.image`, `pystiche.misc` and so on. I've planned some significant changes (see for example pmeier/pystiche#382) that will make parts of this obsolete.

Is there anything I've missed that needs to be addressed?

@edgarriba If possible, I like to know where you see an opening for better error messages. Did you encounter an error where you didn't know how to proceed?
