### 2020-09-24 pmeier/pystiche#401 @pmeier
add citation information to README
As suggested by @NickleDave in https://github.com/pyOpenSci/software-review/issues/25#issuecomment-698361690

Currently we only have a Zenodo badge with a DOI, but the README is lacking a citation information. It should be something along the lines of "If you use this software please cite it as" and provide a BibTex citation.
