### 2020-09-24 pmeier/pystiche#402 @pmeier
add small usage snippet to the README
As suggested by @NickleDave in https://github.com/pyOpenSci/software-review/issues/25#issuecomment-698361690

Currently we only include a link to the examples in the documentation. I was hesitant to add a usage example due to the sheer length. I think it is OK the only include a small snippet showing only the perceptual loss as it is the most important part with a link to the full example.
