### 2020-09-24 pmeier/pystiche#403 @pmeier
Fix installation instructions
Both @edgarriba and @soumith had troubles installing / setting up the development environment in their reviews (see https://github.com/pyOpenSci/software-review/issues/25#issuecomment-695366229 and https://github.com/pyOpenSci/software-review/issues/25#issuecomment-698135408) this was partially resolved in #400, but needs further attention.
