### 2020-09-24 pmeier/pystiche#405 @pmeier
Add inline examples to pystiche.ops and pystiche.ops.functional
As suggested by @soumith in https://github.com/pyOpenSci/software-review/issues/25#issuecomment-698135408

The core functionality of an NST resides in `pystiche.ops`. Thus, I agree with @soumith that it is a good idea to have some inline examples illustrating their usage.
