### 2020-09-24 #25 @pmeier
@NickleDave, both @edgarriba and @soumith noticed that the link to the "pyOpenSi package guidelines" (https://www.pyopensci.org/dev_guide/authoring/overview.html) is broken. Could you provide a working one in order for this piece to also be reviewed?
