### 2020-09-24 00f8a53
fix pyimagetest requirement for tests
