+++ tox.ini
@@ -11,7 +11,7 @@ isolated_build = True
 [testenv:tests]
 deps =
   pytest >= 6
-  pyimagetest
+  pyimagetest < 0.3
   pillow_affine
   dill
   pytest-subtests
