### 2020-09-24 e834eec
fix installation instructions (#406)
