### 2020-09-24 dae5d8f
run lint-style on every PR
