### 2020-09-24 132a11c
run lint-style on every PR (#407)
* run lint-style on every PR

* add newline at EOF
