### 2020-09-24 91d27a4
add citation information to README
