### 2020-09-24 #25 @NickleDave
Yes, I'm sorry about the broken link. I know it needs to be fixed but haven't had a chance to make the PR.
This is the correct link:
https://www.pyopensci.org/contributing-guide/authoring/overview.html

But I do not think it's necessary for @edgarriba or @soumith to check whether you have met the basic guidelines.
Please don't worry about that holding up the review.
The guidelines overlap with what is in the review checklist, and all of them are criteria that you had met at least since submitting if not much earlier. The link is basically there as a reminder to reviewers of what's expected from authors.
