### 2020-09-25 fe2e9d7
additionally run tests on a daily schedule
