+++ .github/workflows/tests.yml
@@ -20,6 +20,9 @@ on:
       - "requirements-dev.txt"
       - ".github/workflows/tests.yml"
 
+  schedule:
+    - cron: "0 4 * * *"
+
 jobs:
   integration:
     strategy:
