+++ tests/integration/demo/test_demo.py
@@ -1,40 +1,41 @@
-import itertools
-from os import path
-
 import pytest
 
-from torchvision.datasets.utils import calculate_md5
-
-from pystiche import demo
+from pystiche import data, demo
 
 from tests import asserts
 
 
-@pytest.mark.large_download
-@pytest.mark.slow
-def test_demo_images_smoke(subtests, tmpdir):
+def test_demo_images_smoke():
     images = demo.demo_images()
-    images.download(root=tmpdir)
-    images = [image for _, image in images]
+    assert isinstance(images, data.DownloadableImageCollection)
 
-    guides = [
-        [guide for _, guide in image.guides]
-        for image in images
-        if image.guides is not None
-    ]
+    names, _ = zip(*iter(images))
+    assert set(names) == {
+        "bird1",
+        "paint",
+        "bird2",
+        "mosaic",
+        "castle",
+        "church",
+        "cliff",
+    }
 
-    for image_or_guide in itertools.chain(images, *guides):
-        with subtests.test(image_or_guide=image_or_guide):
-            file = path.join(tmpdir, image_or_guide.file)
 
-            assert path.exists(file), f"File {file} does not exist."
+@pytest.mark.parametrize(
+    ("name", "regions"),
+    [
+        ("castle", ("building", "water", "sky")),
+        ("church", ("building", "sky")),
+        ("cliff", ("water",)),
+    ],
+)
+def test_demo_images_guides_smoke(name, regions):
+    image = demo.demo_images()[name]
 
-            actual = calculate_md5(file)
-            desired = image_or_guide.md5
-            assert actual == desired, (
-                f"The actual and desired MD5 hash of the image mismatch: "
-                f"{actual} != {desired}"
-            )
+    assert isinstance(image.guides, data.DownloadableImageCollection)
+
+    names, _ = zip(*iter(image.guides))
+    assert set(names).issuperset(set(regions))
 
 
 def test_demo_logger_smoke(caplog):
