+++ tox.ini
@@ -47,7 +47,7 @@ commands =
     tests/galleries \
     {posargs}
 
-[testenv:tests-hub]
+[testenv:tests-download]
 passenv =
   PYSTICHE_HUB_GITHUB
   GITHUB_ACTIONS
@@ -56,7 +56,7 @@ deps = {[tests-common]deps}
 force_cpu = True
 commands =
   {[tests-common]commands} \
-    tests/hub \
+    tests/download \
     {posargs}
 
 [testenv:lint-style]
