### 2020-09-25 5d58b9c
create dedicated download tests workflow
