### 2020-09-25 f7ac254
dont run large download tests on PR triggers
