### 2020-09-25 4d7d520
Run test suite on a schedule (#409)
* additionally run tests on a daily schedule

* merge download tests

* create dedicated download tests workflow

* dont run large download tests on PR triggers

* rename workflow

* fix workflow

* fix caching
