### 2020-09-26 896747e
robustify download_file
