### 2020-09-26 2d115d5
fix checksums of demo images from Pixabay
