### 2020-09-27 dd949dc
robustify download_file (#410)
* robustify download_file

* deprecate user_agent
