+++ .github/workflows/tests.yml
@@ -96,7 +96,7 @@ jobs:
       - name: Get pystiche.home() directory
         id: pystiche-home
         run: |
-          PYTHON=".tox/tests-download/bin/python"
+          PYTHON=".tox/tests-galleries/bin/python"
           DIR="$($PYTHON -c 'import pystiche; print(pystiche.home())')"
           echo "::set-output name=dir::$DIR"
 
