### 2020-09-28 b3ced67
dont delete demo images after test
