### 2020-09-28 7334082
Revert "dont delete demo images after test"
This reverts commit b3ced670096f8443b8cfede4e996283feb426750.
