### 2020-09-28 fceda86
dont check md5 after download for debugging
