### 2020-09-28 4e109b7
Revert "dont check md5 after download for debugging"
This reverts commit fceda860ae4531ade822589d6721e55612150f9e.
