+++ pystiche/misc.py
@@ -254,8 +254,8 @@ def download_file(
         with open(file, "wb") as fh:
             fh.write(response.read())
 
-    # if md5 is not None and not check_md5(file, md5):
-    #     raise RuntimeError(f"The MD5 checksum of {file} mismatches.")
+    if md5 is not None and not check_md5(file, md5):
+        raise RuntimeError(f"The MD5 checksum of {file} mismatches.")
 
     return file
 
