### 2020-09-28 bab16c2
Revert "add debug output"
This reverts commit 49125f30d7d8b15c933c2a265e76ad62716878e1.
