### 2020-09-28 fccda32
fix checksums of demo images from Pixabay (#411)
* fix checksums of demo images from Pixabay

* add image caching

* add debug output

* fix caching

* dont delete demo images after test

* upload images after failure

* dont check md5 after download for debugging

* host Pixabay images ourselves

* Revert "dont check md5 after download for debugging"

This reverts commit fceda860ae4531ade822589d6721e55612150f9e.

* Revert "dont delete demo images after test"

This reverts commit b3ced670096f8443b8cfede4e996283feb426750.

* Revert "add debug output"

This reverts commit 49125f30d7d8b15c933c2a2
[…]
