### 2020-09-28 f545a38
fix download_file in case of a HTTP error
