### 2020-09-29 4993d71
add usage snippet to README
