### 2020-09-29 5fae646
remove double namespacing from pystiche.demo
