### 2020-09-29 cdc225f
add inline examples to comparison and regularization ops
