### 2020-09-30 7eca690
Remove double namespacing from pystiche.demo (#414)
* remove double namespacing from pystiche.demo

* fix examples

* fix docs

* fix type hints
