+++ docs/source/api/pystiche.demo.rst
@@ -3,3 +3,5 @@
 
 .. automodule:: pystiche.demo
   :members:
+    images,
+    logger
