+++ .github/workflows/tests.yml
@@ -10,6 +10,7 @@ on:
     paths:
       - "pystiche/**"
       - "tests/**"
+      - "examples/**.py"
       - "!tests/download/**"
       - "setup.cfg"
       - "pyproject.toml"
