### 2020-09-30 38f77ff
add doctest to CI
