+++ tox.ini
@@ -114,6 +114,18 @@ changedir = {[docs-common]changedir}
 commands =
   sphinx-build -M latexpdf source build
 
+[testenv:tests-docs]
+passenv =
+  {[docs-common]passenv}
+setenv =
+  PYSTICHE_PLOT_GALLERY = False
+  PYSTICHE_DOWNLOAD_GALLERY = False
+force_cpu = True
+deps = {[docs-common]deps}
+changedir = {[docs-common]changedir}
+commands =
+  sphinx-build -b doctest source build
+
 [testenv:publishable]
 whitelist_externals =
   rm
