### 2020-09-30 34d08b5
use subpackage import in examples (#416)
* use subpackage import in examples

* add changes to examples as workflow trigger
