### 2020-09-30 830d2e4
add doctest to CI (#417)
