### 2020-10-02 7f3bc62
add citation information to README (#408)
* add citation information to README

* fix DOI
