### 2020-10-02 ff0b31d
enable global imports for doctest
