### 2020-10-05 73b2442
add examples to operator containers
