### 2020-10-05 65c4820
add examples to pystiche.ops.functional
