+++ docs/source/conf.py
@@ -195,6 +195,7 @@ import torch
 from torch import nn
 
 from pystiche import enc, ops, loss
+import pystiche.ops.functional as F
 """
 
 # -- Options for HTML output -----------------------------------------------------------
