### 2020-10-05 513066a
Add inline examples to core functionalities (#415)
* add inline examples to comparison and regularization ops

* enable global imports for doctest

* fix examples

* add examples to operator containers

* add examples to pystiche.ops.functional
