+++ pystiche/ops/regularization.py
@@ -29,6 +29,12 @@ class TotalVariationOperator(PixelRegularizationOperator):
             results. Defaults to ``2.0``.
         score_weight: Score weight of the operator. Defaults to ``1.0``.
 
+    Examples:
+
+        >>> op = ops.TotalVariationOperator()
+        >>> input = torch.rand(2, 3, 256, 256)
+        >>> score = op(input)
+
     .. seealso::
 
         The total variation loss was introduced by Mahendran and Vedaldi in
