### 2020-10-05 #25 @pmeier
@edgarriba @soumith @NickleDave

I've addressed the reviews according to https://github.com/pyOpenSci/software-review/issues/25#issuecomment-698473702.
