### 2020-10-05 #25 @NickleDave
@edgarriba and @soumith can you just confirm with a comment whether you feel @pmeier has addressed concerns raised in your reviews (if you weren't going to already)?
