### 2020-10-05 #25 @soumith
I've been following @pmeier 's patches and commits.

I think he addressed all my concerns
