### 2020-10-07 #25 @edgarriba
same here. I believe latest commits cover the mentioned issues.
cool vignette. Pretty clear image telling what's going on ;)
