### 2020-10-08 #25 @NickleDave
Awesome!
Thank you so much @soumith and @edgarriba for finding time to review this!
Since PyOpenSci is still getting off the ground it really means a lot to us that you could contribute.
And I know it has helped @pmeier make real improvements to pystiche (even though it was already quite well developed upon submission)
