### 2020-10-08 #25 @NickleDave
🎉 `pystiche` has been approved by pyOpenSci! Thank you again @pmeier  for submitting `pystiche` and many thanks to @soumith @edgarriba for reviewing this package! 😸
@lwasser tagging you in just to let you know I'm moving forward with this

@pmeier can you please do two more things to wrap up:
- [ ] Add the badge for pyOpenSci peer-review to the README.md of `pystiche`. The badge should be `[![pyOpenSci](https://tinyurl.com/y22nb8up)](https://github.com/pyOpenSci/software-review/issues/issue-number)`
- [ ] Add `pystiche` to the pyOpenSci website. Please open a pr to update [this file](https://github.com/pyOpenSci/pyopensci.github.io/blob/master/_data/packages.yml): to add your package and name to the list of contributors

@pmeier @soumith @edgarriba please let you know if you would be okay with me adding you to our website as contributors.
I will submit a PR myself to add you to [this file](https://github.com/pyOpenSci/pyopensci.github.io/blob/master/_data/contributors.yml) if you are okay with it. I don't want to ask you to submit any more PRs when you've already contributed!

For JOSS review: @pmeier I think you have already done the first two items below, right?
I'm going to tag in @arfon just as a heads up but you should be ready to complete the last item.
- [ ] Submit to JOSS using the Zenodo DOI. We will tag it for expedited review.

All, please feel free to share any feedback about the review process here. The broken links in the documentation are duly noted! I will raise issues for them. Any other comments on documentation or on anything else is welcome and appreciated
