### 2020-10-08 #25 @pmeier
@edgarriba @soumith

Thanks again also from my side! I really appreciate the time you voluntarily spent to help me out.

@NickleDave

- I'm fine with being added as contributor
- Yes, the first two items on the JOSS to do list are already done

Overall, I'm happy with the review process. Here are some things that IMO can be improved:

- My initial comment starting the review process still contains a lot of "TBD"s
- This issue still has its initial labels
- I've mentioned this before, but I would have appreciated a more transparent timeline early on
