### 2020-10-08 dc974da
update installation instructions
