+++ README.rst
@@ -52,7 +52,7 @@ release can be installed with
 
 .. code-block:: sh
 
-  pip install git+https://github.com/pmeier/pystiche@v0.5.1
+  pip install pystiche
 
 Usage
 =====
