### 2020-10-08 724cf43
update installation instructions (#420)
