### 2020-10-08 32dd998
update citation info (#421)
