### 2020-10-09 #25 @NickleDave
> @NickleDave no prob from my side. Nice xperience! open to more colab

Excellent thank you @edgarriba

> @NickleDave
>
>     * I'm fine with being added as contributor
Great

>     * Yes, the first two items on the JOSS to do list are already done
>
>     * I've submitted to JOSS, but I think until the review starts there is no publicly accessible URL to link to

Got it

> Overall, I'm happy with the review process. Here are some things that IMO can be improved:
>
>     * My initial comment starting the review process still contains a lot of "TBD"s. I'm not sure if that is my responsibility to fill or yours. If its mine, it should be stated somewhere
>
>     * This issue still has its initial labels

Yes, my fault. Thank you for pointing it out -- will be aware of this in future reviews

>     * I've mentioned this before, but I would have appreciated a more transparent timeline early on

Understood. @lwasser @choldgraf @mbjoseph @jlpalomino  @leouieda @lheagy this is probably a point we should discuss now that we've finished up this review and can reassess how to move PyOpenSci forward
