### 2020-10-20 #25 @pmeier
@NickleDave @lwasser FYI: JOSS review was successful (openjournals/joss-reviews#2761) and the paper is already [published](https://joss.theoj.org/papers/10.21105/joss.02761).
