### 2020-10-20 #25 @lwasser
heck. yea!! thank you @pmeier and @NickleDave  !! this is awesome. Dave, i should give you superpowers to also retweet from pyopensci if you are open to it!! @choldgraf JOSS and @arfon are super fast once the review goes through POS. it's  fantastic!
