### 2021-01-11 #25 @lwasser
hi  there @NickleDave and  @pmeier i'm just popping back here to confirm what version of pystiche  was accepted? i'm going through logs and noticed at the  very top of this review editors and reviewers and version  weren't filled in! minor but important for record keeping! what version did we accept? Many thanks all!!
