### 2021-01-11 #25 @pmeier
`pystiche==0.6.2`
