### 2021-01-13 #25 @pmeier
@lwasser I need to correct myself:

[`pystiche==0.6.0`](https://github.com/pmeier/pystiche/releases/tag/v0.6.0) was accepted by pyOpenSci whereas [`pystiche==0.6.2`](https://github.com/pmeier/pystiche/releases/tag/v0.6.2) was accepted and published by JOSS. I've edited the top post accordingly.
