### 2022-09-28 #25 @lwasser
hey there @soumith @edgarriba 👋 Just a friendly reminder to take 5-10 minutes to fill out our survey . We really appreciate it.  Thank you in advance for helping us by filling out the survey!! 🙌

✨ Phillip thank you so much for taking the time to fill it out 🙌

[🔗 HERE IS THE SURVEY LINK 🔗 ](https://forms.gle/ZZXcD1hr18tYmY2p6)
